% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{b}
\alias{b}
\title{Variance estimator free of joint inclusion probability calculations for unequal probability sampling, Hajek}
\usage{
b(n, y, pi_i_values)
}
\arguments{
\item{n}{sample size (an integer value).}

\item{y}{need description}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}
}
\description{
Variance estimator free of joint inclusion probability calculations for unequal probability sampling, Hajek
}
\references{
\insertRef{berger2005variance}{ACSampling}
}
