% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CactusRealizationSummary.R
\name{CactusRealizationSummary}
\alias{CactusRealizationSummary}
\title{Cactus Realization Summary Data}
\format{
A data frame with 400 rows and 5 variables
}
\description{
Realizations created from my own data.
}
\details{
\itemize{
  \item variable List variables here
}
}
\examples{
# include example here
}
\references{
\insertRef{thompson1990adaptive}{ACSampling}
\insertRef{saubyadaptive}{ACSampling}
}
