% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopSummaryStats.R
\name{calcPopSummaryStats}
\alias{calcPopSummaryStats}
\title{Calculate Summary Statistics for Patch Population Data}
\usage{
calcPopSummaryStats(popdata, summaryvar, rvar = NULL, popgroupvar,
  spatweights = "S", nrow, ncol)
}
\arguments{
\item{summaryvar}{Vector of variables for which summary statistics should be calculated.}

\item{rvar}{Vector of variables for which ratio estimators should be used.}

\item{popgroupvar}{String identifying the categorical variable identifying the different populations.}

\item{spatweights}{Vector of spatial weight matrix styles. Can take on values "W", "B", "C", "U", "S", and "minmax". See nb2listw for more details.}

\item{population_data}{Data on multiple realizations of patches of the species of interest within the grid of locations (created by \code{createSpeciesPatchPopulations} function).}
}
\value{
Dataframe including summary statistics for each column identified in \code{summaryvar} and for each category identified in \code{popgroupvar}.
}
\description{
Calculates summary statistics for patch population data.
}
\examples{
library(magrittr)
library(dplyr)
ovar = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)		
summaryvar = ovar
# WHAT WAS I THINK HERE? for grouping variables?
popgroupvar = "n.networks" # c("n.networks", "realization")
# create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
start.seed=1
buffer=5
rvar = c("CACA_on_Stricta", "CACA_on_Pusilla")
cactus.realizations <- createSpeciesPatchRealizations(x_start, x_end,
	y_start, y_end, buffer, n.networks, n.realizations, SpeciesInfo, start.seed,
	ovar)
patch_data_summary <- calcPopSummaryStats(cactus.realizations, 
	summaryvar=ovar, popgroupvar=popgroupvar, nrow=30, ncol=30)
}
