% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRE.R
\name{calcRE}
\alias{calcRE}
\title{Calculate Relative Efficiency (RE)}
\usage{
calcRE(
  MSE_ComparisonSamplingDesign,
  popdata,
  popgroupvar,
  samplesizevar,
  rvar,
  ovar
)
}
\arguments{
\item{MSE_ComparisonSamplingDesign}{Sampling design for which relative efficiency (RE) should be calculated.}

\item{popdata}{Dataframe of population data.}

\item{popgroupvar}{Categorical variables identifying the patch realization from which the simulation data was generated (e.g., \code{n.networks} and \code{realization}). WHAT ELSE}

\item{samplesizevar}{Name of column in population data (?) containing the variable indicating variation in sample size.}

\item{rvar}{Ratio variables.}

\item{ovar}{Occupancy variables.}
}
\value{
Dataframe including original data and RE estimates.
}
\description{
Calculate efficiency of sampling design, relative to WHAT.
}
\examples{
# dimensions of populations
dimensions <- popdata \%>\% 
group_by_(.dots=popgroupvar) \%>\%
summarise(N = n())
# variances of populations
CactusRealizationSummary <- calculatePopSummaryStats(
popdata = CactusRealizations, 
summaryvar = c("Stricta", "Pusilla", "Cactus",
	"MEPR_on_Stricta", "CACA_on_Stricta", "Percent_Cover_Stricta", 
	"Height_Stricta", "Old_Moth_Evidence_Stricta"), 
popgroupvar = "population", 
rvar = c("MEPR_on_Stricta", "CACA_on_Stricta", 
	"Percent_Cover_Stricta", "Height_Stricta", 
	"Old_Moth_Evidence_Stricta"),
nrow=30,
ncol=30
)
patch_data_summary_wide <- createWidePopSummaryStats(
popsummarystats = CactusRealizationSummary,
ovar = "Stricta",
rvar = c("MEPR_on_Stricta", "CACA_on_Stricta", "Percent_Cover_Stricta", 
	"Height_Stricta", "Old_Moth_Evidence_Stricta")
)
simulation_data_summary_table_re = calcSamplingBias(
population_data_summary	= patch_data_summary_wide, 
simulation_data		= simdata_all_re, 
sampgroupvar	= sampgroupvar, 
popgroupvar = popgroupvar,
ovar			= ovar, 
rvar				= rvar
)
RE_values <- calcRE(
population_data = patch_data_summary_wide,
MSE_ComparisonSamplingDesign = simulation_data_summary_table_re,
popgroupvar = "population",
samplesizevar = "N.Total.plots_mean",
ovar = ovar,
rvar = rvar
)
}
