% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSamplingBias.R
\name{calcSamplingBias}
\alias{calcSamplingBias}
\title{Calculate Simulation Data Sampling Bias}
\usage{
calcSamplingBias(
  popdatasummary,
  simdata,
  popgroupvar,
  samplinggroupvar,
  ovar,
  orvar,
  rvar
)
}
\arguments{
\item{popdatasummary}{Summary statistics on the species patch realizations of patches (created by \code{calculateRealizationSummaryStatistics} function).}

\item{simdata}{Simulation data on sampling of the multiple patch realizations.}

\item{popgroupvar}{Categorical variable used to identify different populations.}

\item{samplinggroupvar}{Categorical variable used to identify different samples.}

\item{ovar}{Vector of occupancy variables. The total vector of variables (\code{c(avar, ovar, rvar)}) should be a length of at least 1.}

\item{orvar}{Vector of variables for which secondary variables should be estimated. Can be identical to \code{ovar} or a subset.}

\item{rvar}{Vector of ratio variables. The total vector of variables (\code{c(avar, ovar, rvar)}) should be a length of at least 1.}
}
\value{
Dataframe including simulation data summary statistics, including relative bias and mean squared error (MSE) of the mean and variance.
}
\description{
Calculate the sampling bias of different sampling designs from simulation data.
}
\examples{
# Create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
Species.Fields = c("Stricta", "Pusilla", "Cactus")
cactus.realizations = createRealizations(x_start, x_end,
	y_start, y_end, buffer=5, n.networks, n.realizations, SpeciesInfo, 
	start.seed=1, Species.Fields, yvar="Cactus")

# Sample from the realizations
simulations=1
nsamples=c(5,10,20,40)
population <- createPop(x_start = 1, x_end = 30, y_start = 1, 
	y_end = 30)
abundance.variables = NULL
occupancy.variables = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)		
patch_data = cactus.realizations
# simdata <- sampleSpeciesPatchRealizations(patch_data, simulations, 
#	nsamples, population, abundance.variables, occupancy.variables)

# summary.variables = occupancy.variables
# grouping.variables = c("n.networks", "realization")
# dataset = cactus.realizations
# patch_data_summary <- calculateRealizationSummaryStatistics(dataset, 
# 	summary.variables, grouping.variables)
sims=200
n1=c(5,10,20,40)
population <- createPop(x_start = 1, x_end = 30, y_start = 1, y_end = 30)
avar = NULL
ovar = c(
"Stricta",
"Pusilla",
"Cactus",
"CACA_on_Pusilla",
"CACA_on_Stricta",
"MEPR_on_Pusilla",
"MEPR_on_Stricta",
"Old_Moth_Evidence_Pusilla",
"Old_Moth_Evidence_Stricta"
# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
# "Percent_Cover_Stricta",
# "Height_Pusilla",
# "Height_Stricta",
)		
popdata = cactus.realizations
#simulation_data <- sampleRealizations(popdata, sims, 
#n1, population, avar, ovar)
#CactusRealizationSummary <- calcPopSummaryStats(
#	popdata = CactusRealizations, 
#	summaryvar = c("Stricta", "Pusilla", "Cactus",
#		"MEPR_on_Stricta", "CACA_on_Stricta", "Percent_Cover_Stricta", 
#		"Height_Stricta", "Old_Moth_Evidence_Stricta"), 
#	popgroupvar = "population", 
#	rvar = c("MEPR_on_Stricta", "CACA_on_Stricta", 
#		"Percent_Cover_Stricta", "Height_Stricta", 
#		"Old_Moth_Evidence_Stricta"),
#	nrow=30,
#	ncol=30
#)
#patch_data_summary_wide <- createWidePopSummaryStats(
#	popsummarystats = CactusRealizationSummary,
#	ovar = "Stricta",
#	rvar = c("MEPR_on_Stricta", "CACA_on_Stricta", "Percent_Cover_Stricta", 
#		"Height_Stricta", "Old_Moth_Evidence_Stricta")
#)
#simdata_summary_table_re = calcSamplingBias(
#	popdatasummary	= patch_data_summary_wide, 
#	simdata		= simdata_all_re, 
#	sampgroupvar	= sampgroupvar, 
#	popgroupvar = popgroupvar,
#	ovar			= ovar, 
#	rvar				= rvar
#)
# avar = NULL
}
