% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations_depends_estimators.R
\name{calcSpatStats}
\alias{calcSpatStats}
\title{Calculate Spatial Statistics}
\usage{
calcSpatStats(alldata_all, weights, yvar)
}
\arguments{
\item{alldata_all}{NEED TO DESCRIBE.}

\item{weights}{If SpatStat is "TRUE", this is a vector giving 
spatial weight matrix styles to use to calculate the Join Count and Moran's I
 statistics. Can take on values "W", "B", "C", "U", "S", and "minmax". See 
 \code{nb2listw} for more details.}
}
\description{
Calculate Spatial Statistics
}
