% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSSQR.R
\name{calcSSQR}
\alias{calcSSQR}
\title{Calculate SSQ_R}
\usage{
calcSSQR(popdata, variable, popgroupvar = NA)
}
\arguments{
\item{popdata}{Information about the populations of interest.}

\item{variable}{Variable for which to calculate SSQr.}

\item{popgroupvar}{A variable identifying the separate populations. If only one population in popdata, set popgroupvar = NULL.}
}
\value{
Dataframe including original data and RE estimates.
}
\description{
SSQ_R is the ratio of the within-network sum of squares to the total sum of squares. ACS design becomes more efficient relative to simple random sampling as the within-network variation increases relative to the overall variation [@thompson1996adaptive]. We calculate \deqn{SSQ_R = SSQ_w/SSQ_\tau}{SSQ_R = SSQ_w/SSQ_tau} as the ratio of the within-network sum of squares \deqn{SSQ_w = \sum_{j=1}^{\kappa} \sum_{i \in j} (y_{j,i} - \bar{y_{j}})^2}{SSQ_w = \sum_{j=1}^{\kappa} \sum_{i \in j} (y_{j,i} - \bar{y_{j}})^2}, to the total sum of squares \deqn{SSQ_\tau= \sum_{i=1}^{N} (y_i - \mu)^2}{SSQ_\tau= \sum_{i=1}^{N} (y_i - \mu)^2}. Thus, an increase in \code{SSQ_R} indicates an increase in the efficiency of the ACS design relative to SRSWOR.
}
\examples{
data(Thompson1990Figure1Population)
temp <- Thompson1990Figure1Population \%>\%
	mutate(pop = 1)
popdata <- temp
variable <- "y_value"
popgroupvar <- "pop"
calcSSQR(popdata, variable, popgroupvar)
popgroupvar <- NA
calcSSQR(popdata, variable, popgroupvar)
}
\references{
su2003estimator
}
