% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateBiasOfVariance.R
\name{calculateBiasOfVariance}
\alias{calculateBiasOfVariance}
\title{calculate bias of the variance}
\usage{
calculateBiasOfVariance(dataset, patch_data_summary, variables,
  Sampling = "ACS")
}
\arguments{
\item{dataset}{Dataset of simluation results}

\item{patch_data_summary}{Dataset summarizing known true parameters of the habitat patches sampled in the simulations}

\item{variables}{Variables for which to calculate bias}

\item{Sampling}{Defaults to "ACS"}
}
\description{
calculate relative bias of the variance
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}

