% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJointInclusionProbabilities.R
\name{calculateJointInclusionProbabilities}
\alias{calculateJointInclusionProbabilities}
\title{Calculate Joint Inclusion Probabilities Using Simulations}
\usage{
calculateJointInclusionProbabilities(patchdat, simulations, nsamples,
  SamplingDesign = "ACS", y_variable, f_max = NULL,
  population.grouping.variable = NULL)
}
\arguments{
\item{patchdat}{A data frame specifying the patch realizations. WHAT FORMAT SHOULD IT BE IN?}

\item{simulations}{Number of simulations per population.}

\item{nsamples}{Vector of initial sample size(s) for the initial simple random sample(s) without replacement; can be a single value or vector of values}

\item{SamplingDesign}{Sampling design; ACS or RACS. Default value is "ACS."}

\item{y_variable}{variable upon which adaptive cluster sampling criterion is based}

\item{f_max}{Default value is NULL.}

\item{population.grouping.variable}{variable identifying unique populations. Default value is NULL.}
}
\value{
A list of nsamples and dat, BUT I DONT REMEMBER WHAT THOSE ARE.
}
\description{
Calculate inclusion probabilities for each unit in a population using simulations.
}
