% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSamplingBias.R
\name{calculateMSE}
\alias{calculateMSE}
\title{Calculate Mean Squared Error}
\usage{
calculateMSE(simulation_data_summary, grouping.variables, variables, rvar,
  statistics)
}
\arguments{
\item{simulation_data_summary}{Calculate Mean Squared Error (MSE).}

\item{grouping.variables}{Categorical variables identifying the patch realization from which the simulation data was generated (e.g., \code{n.networks} and \code{realization}).}

\item{variables}{Vector of variables for which MSE should be estimated.}

\item{rvar}{Variables for which to use ratio estimators}

\item{statistics}{Statistics (e.g., mean and variance) for which to calculate MSE}
}
\value{
Dataframe including original data and MSE estimates.
}
\description{
Calculate Mean Squared Error
}

