% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopulationSummaryStatistics.R
\name{calculatePopulationSummaryStatistics}
\alias{calculatePopulationSummaryStatistics}
\title{Calculate Summary Statistics for Patch Population Data}
\usage{
calculatePopulationSummaryStatistics(population_data, summary.variables,
  ratio.variables = NULL, population.grouping.variable, weights = "S")
}
\arguments{
\item{population_data}{Data on multiple realizations of patches of the species of interest within the grid of locations (created by \code{createSpeciesPatchPopulations} function).}

\item{summary.variables}{Vector of variables for which summary statistics should be calculated.}

\item{ratio.variables}{Variables for which to use ratio estimators}

\item{population.grouping.variable}{Categorical variable identifying the different populations.}

\item{weights}{Vector of spatial weight matrix styles. Can take on values "W", "B", "C", "U", "S", and "minmax". See nb2listw for more details.}
}
\value{
Dataframe including summary statistics for each column identified in \code{summary.variables} and for each category identified in \code{grouping.variables}.
}
\description{
Calculates summary statistics for patch population data.
}
\examples{
library(magrittr)
library(dplyr)
occupancy.variables = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)		
summary.variables = occupancy.variables
# WHAT WAS I THINK HERE? for grouping variables?
grouping.variables = "n.networks" # c("n.networks", "realization")
# create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
start.seed=1
buffer=5
cactus.realizations <- createSpeciesPatchRealizations(x_start, x_end,
	y_start, y_end, buffer, n.networks, n.realizations, SpeciesInfo, start.seed,
	occupancy.variables)
patch_data_summary <- calculatePopulationSummaryStatistics(cactus.realizations, 
	summary.variables=occupancy.variables, population.grouping.variable=grouping.variables)
}
