% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRelativeEfficiency.R
\name{calculateRE}
\alias{calculateRE}
\title{Calculate Relative Efficiency (RE)}
\usage{
calculateRE(MSE_ComparisonSamplingDesign, population_data,
  population.grouping.variables, sample.size.variable, rvar, ovar)
}
\arguments{
\item{MSE_ComparisonSamplingDesign}{Sampling design for which relative efficiency (RE) should be calculated.}

\item{population_data}{Dataframe of population data.}

\item{population.grouping.variables}{Categorical variables identifying the patch realization from which the simulation data was generated (e.g., \code{n.networks} and \code{realization}). WHAT ELSE}

\item{sample.size.variable}{Name of column in population data (?) containing the variable indicating variation in sample size.}

\item{rvar}{Ratio variables.}

\item{ovar}{Occupancy variables.}
}
\value{
Dataframe including original data and RE estimates.
}
\description{
Calculate efficiency of sampling design, relative to WHAT.
}
