% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRealizationSummaryStatistics.R
\name{calculateRealizationSummaryStatistics}
\alias{calculateRealizationSummaryStatistics}
\title{Calculate Summary Statistics for Patch Realization Data}
\usage{
calculateRealizationSummaryStatistics(dataset, summary.variables,
  ratio.variables = NULL, grouping.variables)
}
\arguments{
\item{dataset}{Data on multiple realizations of patches of the species of interest within the grid of locations (created by \code{createSpeciesPatchRealizations} function).}

\item{summary.variables}{Vector of variables for which summary statistics should be calculated.}

\item{ratio.variables}{Variables for which to use ratio estimators}

\item{grouping.variables}{Categorical variables identifying the patch realization from which the simulation data was generated (e.g., \code{n.networks} and \code{realization}).}
}
\value{
Dataframe including summary statistics for each column identified in \code{summary.variables} and for each category identified in \code{grouping.variables}.
}
\description{
Calculate Summary Statistics for Patch Realization Data
}
\examples{
library(magrittr)
library(dplyr)
occupancy.variables = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)		
summary.variables = occupancy.variables
grouping.variables = c("n.networks", "realization")
# create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
start.seed=1
buffer=5
cactus.realizations <- createSpeciesPatchRealizations(x_start, x_end,
	y_start, y_end, buffer, n.networks, n.realizations, SpeciesInfo, start.seed,
	occupancy.variables)
patch_data_summary <- calculateRealizationSummaryStatistics(cactus.realizations, 
	summary.variables=occupancy.variables, grouping.variables=grouping.variables)
patch_data_summary \%<>\% 
	round(3) \%>\% 
	arrange(n.networks) \%>\% 
	dplyr::select(
		starts_with("Pusilla"), 
		starts_with("Stricta"), 
		starts_with("Cactus"), 
		starts_with("CACA_on_Pusilla"),
		starts_with("CACA_on_Stricta"),
		starts_with("MEPR_on_Pusilla"),
		starts_with("MEPR_on_Stricta"),
		starts_with("Old_Moth_Evidence_Pusilla"),
		starts_with("Old_Moth_Evidence_Stricta"),
		everything()
	)
write.csv(patch_data_summary, file=paste("patch_data_summary", 
format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M"), ".csv", sep=""))
}

