% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSamplingBias.R
\name{calculateSamplingBias}
\alias{calculateSamplingBias}
\title{Calculate Simulation Data Sampling Bias}
\usage{
calculateSamplingBias(population_data_summary, simulation_data,
  population.grouping.variables, sampling.grouping.variables, ovar, o_rvar,
  rvar)
}
\arguments{
\item{population_data_summary}{Summary statistics on the species patch realizations of patches (created by \code{calculateRealizationSummaryStatistics} function).}

\item{simulation_data}{Simulation data on sampling of the multiple patch realizations.}

\item{population.grouping.variables}{Categorical variables with which to group the population data (e.g., artificial population number if there are more than 1)}

\item{sampling.grouping.variables}{Categorical variables with which to group the simulation data (e.g., sampling design used, number of primary samples).}

\item{ovar}{Vector of variables for which sampling bias should be estimated.}

\item{o_rvar}{Vector of variables for which secondary variables should be estimated. Can be identical to ovar or a subset.}

\item{rvar}{Variables for which to use ratio estimators}
}
\value{
Dataframe including simulation data summary statistics, including relative bias and mean squared error (MSE) of the mean and variance.
}
\description{
Calculate the sampling bias of different sampling designs from simulation data.
}
\examples{
# Create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
Species.Fields = c("Stricta", "Pusilla", "Cactus")
cactus.realizations = createSpeciesPatchRealizations(x_start, x_end,
	y_start, y_end, buffer=5, n.networks, n.realizations, SpeciesInfo, 
	start.seed=1, Species.Fields)

# Sample from the realizations
simulations=1
nsamples=c(5,10,20,40)
population <- createPopulation(x_start = 1, x_end = 30, y_start = 1, 
	y_end = 30)
abundance.variables = NULL
occupancy.variables = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)		
patch_data = cactus.realizations
# simulation_data <- sampleSpeciesPatchRealizations(patch_data, simulations, 
#	nsamples, population, abundance.variables, occupancy.variables)

# summary.variables = occupancy.variables
# grouping.variables = c("n.networks", "realization")
# dataset = cactus.realizations
# patch_data_summary <- calculateRealizationSummaryStatistics(dataset, 
# 	summary.variables, grouping.variables)
# avar = NULL
}
