% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSSQR.R
\name{calculateSSQR}
\alias{calculateSSQR}
\title{Calculate SSQ_R}
\usage{
calculateSSQR(patch_data, variable, population.grouping.variable)
}
\arguments{
\item{patch_data}{Information about the populations of interesting}

\item{variable}{Variable for which to calculate SSQr?}

\item{population.grouping.variable}{variable identifying the separate populations. If only one population in patch_data, set population.grouping.variable = NULL.}
}
\value{
Dataframe including original data and RE estimates.
}
\description{
SSQ_R is the ratio of the within-network sum of squares to the total sum of squares. ACS design becomes more efficient relative to simple random sampling as the within-network variation increases relative to the overall variation [@thompson1996adaptive]. We calculate $SSQ_R = SSQ_w/SSQ_\tau$ as the ratio of the within-network sum of squares $SSQ_w = \sum_{j=1}^{\kappa} \sum_{i \in j} (y_{j,i} - \bar{y_{j}})^2$, to the total sum of squares $SSQ_\tau= \sum_{i=1}^{N} (y_i - \mu)^2$ [`r table_captions("cactus_summary_cap", display = "cite")`, @su2003estimator]. Thus, an increase in $SSQ_R$ indicates an increase in the efficiency of the ACS design relative to SRSWOR.
}
