% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCactusRealizations.R
\name{createCactusRealizations}
\alias{createCactusRealizations}
\title{Create Cactus Realization Data}
\usage{
createCactusRealizations()
}
\value{
The data from Sauby and Christman.
}
\description{
Create Cactus Realization Data
}
\examples{
library(ggplot2)
CactusRealizations <- createCactusRealizations()
ggplot(
	CactusRealizations,
	aes(
		x, 
		y, 
		shape = factor(Cactus)
)) +
	geom_point() +
	facet_wrap(~population) +
	scale_shape_manual(values=c(4,16))

}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}

