% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCactusRealizationSummary.R
\name{createCactusRealizationSummary}
\alias{createCactusRealizationSummary}
\title{Create Summary of Cactus Realization Data}
\usage{
createCactusRealizationSummary(CactusRealizations)
}
\value{
The data from Sauby and Christman.
}
\description{
Create Summary of Cactus Realization Data
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
