% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlambdap_5_tau_25.R
\name{createlambdap_5_tau_25}
\alias{createlambdap_5_tau_25}
\title{Create the population, \deqn{\lambda_P = 5, \tau = 25}{lambda_P = 5, tau = 25}, from Christman (1997)}
\usage{
createlambdap_5_tau_25()
}
\value{
The population displayed in Christman (1997), where \deqn{\lambda_P = 5, \tau = 25}{lambda_P = 5, tau = 25}.
}
\description{
Create the population, \deqn{\lambda_P = 5, \tau = 25}{lambda_P = 5, tau = 25}, from Christman (1997)
}
\examples{
library(ggplot2)
library(magrittr)
data(lambdap_5_tau_25)

ggplot(lambdap_5_tau_25 \%>\% dplyr::filter(y_value!=0), aes(x, y)) + geom_text(aes(label=y_value))
}
\references{
Christman, M. C. (1997). Efficiency of some sampling designs for spatially clustered populations. \emph{Environmetrics}, 8: 145--166.
}
