% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createNetworkCenters.R
\name{createNetworkCenters}
\alias{createNetworkCenters}
\title{Create Network Centers}
\usage{
createNetworkCenters(Species.Data)
}
\arguments{
\item{Species.Data}{dataframe containing Network IDs, x and y coordinates}
}
\description{
Takes dataframe containing Network IDs, x and y coordinates, and returns the coordinates of the centers of the networks, in integer values.
coordinates should be labeled x and y and Network ID given as "NetworkID"
}
