% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPopulation.R
\name{createPop}
\alias{createPop}
\title{Create a population of locations (in a grid).}
\usage{
createPop(x_start, x_end, y_start, y_end)
}
\arguments{
\item{x_start}{The most western coordinate of the grid.}

\item{x_end}{The most eastern coordinate of the grid.}

\item{y_start}{The most southern coordinate of the grid.}

\item{y_end}{The most northern coordinate of the grid.}
}
\value{
A grid of locations with coordinates (x, y).
}
\description{
Create a population of locations (in a grid).
}
