% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSRSWOR.R
\name{createSRSWOR}
\alias{createSRSWOR}
\title{Create an Adaptive Cluster Sample.}
\usage{
createSRSWOR(population, seed, n1)
}
\arguments{
\item{population}{grid of population to be sampled.}

\item{seed}{vector of numbers to feed to \code{set.seed()} so that the sampling is reproducible.}

\item{n1}{initial sample size (sampled according to simple random sampling without replacement).}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Create an Adaptive Cluster Sample.
}
\examples{
# example
# create the population
# create the patch
# then sample

data(Thompson1990Figure1Population)
data(Thompson1990Figure1Sample)
Z = createSRSWOR(Thompson1990Figure1Population, seed=2, n1=10)
}

