% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createThompson1990Figure1Population.R
\name{createThompson1990Figure1Population}
\alias{createThompson1990Figure1Population}
\title{Create the population displayed in Figure 1 from Thompson (1990)}
\usage{
createThompson1990Figure1Population()
}
\value{
The population displayed in Figure 1 from Thompson (1990).
}
\description{
Create the population displayed in Figure 1 from Thompson (1990)
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)
Thompson1990Figure1Population <- createThompson1990Figure1Population()
p <- ggplot(Thompson1990Figure1Population \%>\% filter(m>1), aes(x, y, 
	colour=NetworkID, label=NetworkID))
p + annotate("rect", xmin=1, xmax=20, ymin=1, ymax=20, alpha=0, 
colour="grey") + 
	geom_point(aes(size=factor(y_value))) + 
	scale_colour_gradientn(colours = rainbow(7))
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

