% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createThompson1990Figure1Sample.R
\name{createThompson1990Figure1Sample}
\alias{createThompson1990Figure1Sample}
\title{Create the population displayed in Figure 1 from Thompson (1990)}
\usage{
createThompson1990Figure1Sample()
}
\value{
The population displayed in Figure 1 from Thompson (1990).
}
\description{
Create the population displayed in Figure 1 from Thompson (1990)
}
\examples{
createThompson1990Figure1Sample()
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

