% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createThompson1990Table1data.R
\name{createThompson1990Table1data}
\alias{createThompson1990Table1data}
\title{Create the Table 1 data from Thompson (1990)}
\usage{
createThompson1990Table1data()
}
\value{
The data displayed in Table 1 from Thompson (1990).
}
\description{
Create the Table 1 data from Thompson (1990)
}
\examples{
createThompson1990Table1data()
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

