% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWidePopulationSummaryStatistics.R
\name{createWidePopulationSummaryStatistics}
\alias{createWidePopulationSummaryStatistics}
\title{Create "Wide" Format Population Summary Statistics for Patch Population Data}
\usage{
createWidePopulationSummaryStatistics(PopulationSummaryStatistics, ovar,
  rvar)
}
\arguments{
\item{PopulationSummaryStatistics}{Created by the function calculatePopulationSummaryStatistics.}

\item{ovar}{list of occupancy variables}

\item{rvar}{list of ratio variables}
}
\value{
Dataframe with a column per mean/variance of each variable and a row per population.
}
\description{
Create "Wide" Format Population Summary Statistics for Patch Population Data
}
