% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations_depends_estimators.R
\name{getJoinCountTestEst}
\alias{getJoinCountTestEst}
\title{Calculate the Join Count Test estimate}
\usage{
getJoinCountTestEst(spatdata, lwb)
}
\arguments{
\item{spatdata}{dataset in which columns x and y have been identified as 
geographic coordinates using the `coordinates` function}

\item{lwb}{a weights list created using the `nb2listw` function}
}
\description{
Calculate the Join Count Test estimate
}
