% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations_depends_estimators.R
\name{getMoranTestEst}
\alias{getMoranTestEst}
\title{Calculate the Moran Test estimate}
\usage{
getMoranTestEst(spatdata, lwb)
}
\arguments{
\item{spatdata}{dataset in which columns x and y have been identified as 
geographic coordinates using the `coordinates` function}

\item{lwb}{a weights list created using the `nb2listw` function}
}
\description{
Calculate the Moran Test estimate
}
