% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{Hajek}
\alias{Hajek}
\title{Hakey variance estimator free of joint inclusion probability calculations for unequal probability sampling}
\usage{
Hajek(pi_i, n)
}
\arguments{
\item{n}{sample size (an integer value).}

\item{pi_i_values}{vector of first-order inclusion probabilities, calculated using \code{pi_i}.}
}
\description{
Option for the "var_pi" function.
}
\references{
@template Hajek_1964
}
