% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChristmanDataExamples.R
\name{lambdap_10_tau_5}
\alias{lambdap_10_tau_5}
\title{Example dataset from Christman (1997)}
\format{
A data frame with 100 rows and 5 variables
}
\description{
An example dataset from Christman (1997): \eqn{\lambda_P = 10, \tau = 5}{lambda_P = 10, tau = 5}.
}
\details{
\itemize{
  \item y. Y coordinate.
  \item x. X coordinate.
  \item NetworkID. The ID of the network to which the unit belongs.
  \item m. The number of units in the given network.
  \item y_value. The abundance of the species of interest.
}
}
\examples{
library(ggplot2)
library(magrittr)
data(lambdap_5_tau_1)

ggplot(lambdap_5_tau_1 \%>\% dplyr::filter(y_value!=0), aes(x, y)) + geom_text(aes(label=y_value))
}
\references{
\insertRef{christman1997efficiency}{ACSampling}
}
