% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopSummaryStats.R
\name{Mean}
\alias{Mean}
\title{Return the mean of a vector with up to two decimal places, after removing NAs.}
\usage{
Mean(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}
