% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCactusRealizations.R
\name{NA_is_Zero_Function}
\alias{NA_is_Zero_Function}
\title{Change NA values to 0.}
\usage{
NA_is_Zero_Function(x)
}
\arguments{
\item{x}{Vector of data.}
}
\description{
Change NA values to 0.
}
