% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_HT.R
\name{new_y_HT}
\alias{new_y_HT}
\title{Calculate the Horvitz-Thompson mean of an adaptive cluster sample, NEW FORMULA.}
\usage{
new_y_HT(
  y,
  N,
  n1,
  m_threshold,
  pi_i_values = NULL,
  m_vec = NULL,
  sampling = NULL,
  criterion = NULL
)
}
\arguments{
\item{y}{The variable of interest, $y$. Must be a numeric vector. The criterion that determines whether adaptive cluster sampling takes place is based on this variable.}

\item{m_threshold}{threshold value above which to calculate pi_i and pi_j differently.}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}

\item{sampling}{A vector (\code{character} format) describing whether units were included in the initial sample or subsequent ACS sample. Units selected in the initial sample should be given the value "Initial_Sample" in the \code{sampling} vector.}

\item{criterion}{Numeric threshold value of \code{yvar} that initiates ACS. Defaults to 0 (ie., anything greater than 0 initiates adaptive cluster sampling).}
}
\value{
The Horvitz-Thompson mean.
}
\description{
This calculate the Horvitz-Thompson mean of an adaptive cluster sample done by sampling without replacement.

where $v$ is the number of distinct units in the sample and
$J_k$ is an indicator variable, equalling 0 if the $k$ th unit in the sample does not satisfy the condition and was not selected in the initial sample; otherwise, $J_k = 1$.
}
\examples{
@template ex_Thompson2002_2_p_307_values
}
\references{

}
