% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupancy_function.R
\name{occupancyfunction}
\alias{occupancyfunction}
\title{Occupancy Function}
\usage{
occupancyfunction(x)
}
\arguments{
\item{x}{vector of presence/absence data}
}
\description{
For multiple observations of the sample plot, determine how many presences were observed.
}
