% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopSummaryStats.R
\docType{data}
\name{ovar}
\alias{ovar}
\title{Calculate Summary Statistics for Patch Population Data}
\format{
An object of class \code{character} of length 13.
}
\usage{
ovar
}
\arguments{
\item{popdata}{Data on multiple realizations of patches of the species of interest within the grid of locations (created by \code{createSpeciesPatchPopulations} function).}

\item{summaryvar}{Vector of variables for which summary statistics should be calculated.}

\item{popgroupvar}{Categorical variable used to identify different populations.}

\item{rvar}{Vector of ratio variables. The total vector of variables (\code{c(avar, ovar, rvar)}) should be a length of at least 1.}

\item{spatweights}{Vector of spatial weight matrix styles. Can take on values "W", "B", "C", "U", "S", and "minmax". See \code{nb2listw} for more details.}

\item{nrow}{the number of rows in the grid that creates the population}

\item{ncol}{the number of columns in the grid that creates the population}
}
\value{
A list of two dataframes, where the first dataframe includes network- and population-specific summary statistics about $m$, summarizes across unique networks as well as unique units within each population, and where the second list includes summary statistics about the \code{summaryvar}.
}
\description{
Calculates summary statistics for patch population data, including summary statistics about $m$ (minimum, maximum, mean, and variance) for each network and population as a whole, and the number of networks with $m>1$.
}
\examples{
library(magrittr)
library(dplyr)
}
\keyword{datasets}
