% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_i.R
\name{pi_i}
\alias{pi_i}
\title{Calculate the inclusion probabilities of units in a simple random sample without replacement.}
\usage{
pi_i(N, n1, m_vec)
}
\value{
vector of inclusion probabilities
}
\description{
Calculate the inclusion probabilities of units in a simple random sample without replacement.
}
\examples{
# example from section 4, p. 1055 (first column) in Thompson (1990)
N = 5
n1 = 2
y = c(1,10,1000)
m_vec = c(1,2,2)
pi_i(N, n1, m)
# EXAMPLE 4:
# Ch. 24, Exercise #2, p. 307, from Thompson (2002)
N=1000
n1=100
m_vec=c(2,3,rep(1,98))
head(pi_i(N, n1, m_vec))
}
