% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ij.R
\name{pi_ij_replace}
\alias{pi_ij_replace}
\title{Calculate joint inclusion probability of unit $j$ and $h$ when sampling occurs with replacement}
\usage{
pi_ij_replace(N, n1, m_vec)
}
\description{
Calculate joint inclusion probability of unit $j$ and $h$ when sampling occurs with replacement
}
