% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ij_with_replacement.R
\name{pi_ij_with_replacement}
\alias{pi_ij_with_replacement}
\title{Calculate joint inclusion probability of unit $j$ and $h$}
\usage{
pi_ij_with_replacement(N, n1, m)
}
\arguments{
\item{N}{Population size.}

\item{n1}{Initial sample size.}

\item{m}{Vector of values giving the number of units satisfying the ACS criterion in network $i$.}
}
\description{
Calculate joint inclusion probability of unit $j$ and $h$
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}

