% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotInfo.R
\name{PlotInfo}
\alias{PlotInfo}
\title{Information about plots surveyed within the Guana Tolomato Matanzas National Estuarine Research Reserve}
\format{A data frame with 1372 rows and 14 variables}
\arguments{
\item{Island}{Habitat patch within which plot is located}

\item{Tag_Number}{Identification number unique to each plot}

\item{Cluster}{The cluster ID to which a plot belongs}

\item{Cluster2}{If a plot belongs to more than one cluster (when clusters overlap), this is used to indicate the second cluster ID to which a plot belongs}

\item{Network}{The network to which a plot belongs}

\item{Sampling}{Whether a plot was included in the initial sample (according to simple random sampling without replacement) or sampled according to restricted adaptive cluster sampling}

\item{Easting}{The Easting coordinate, in UTM}

\item{Northing}{The Northing coordinate, in UTM}

\item{Sample_Number}{For plots sampled according to the initial sample, this is the number indicating the order in which plots were sampled}

\item{PlotDescription}{Notes about plot location or other pertinent information}

\item{SurveyOrder}{The order in which plots should be surveyed so as to minimize traveling time}

\item{OtherPlotNotes}{Other notes about the plot}

\item{RemovedFromStudy}{If yes, the plot was removed from the study because it was not relocated due to either erosion or through burial in developing sand dunes}

\item{PlantsOutsideofPlots}{I can't remember what this was for!}
}
\description{
Information about plots surveyed within the Guana Tolomato Matanzas National Estuarine Research Reserve
}

