% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSurveys_season1.R
\name{PlotSurveys_season1}
\alias{PlotSurveys_season1}
\title{Information about plots surveyed within the Guana Tolomato Matanzas National Estuarine Research Reserve
Survey information for plots containing cacti, surveyed during Summer 2012.
\itemize{
		\item NetworkID Do not use; number of unique Network ID's equals 86, not 88
		\item Island Identifying number/name for habitat patch
		\item unitID The Tag_Number of the plot
		\item n.surveys
		\item Date
		\item Network.mod Tag_Number identifier for each unique Network (n = 88)
		\item x Easting
		\item y Northing
		\item Stricta
		\item Pusilla
		\item Cactus
		\item CACA_on_Pusilla
		\item CACA_on_Stricta
		\item MEPR_on_Pusilla
		\item MEPR_on_Stricta
		\item Old_Moth_Evidence_Pusilla
		\item Old_Moth_Evidence_Stricta
		\item Percent_Cover_Pusilla
		\item Percent_Cover_Stricta
		\item Height_Pusilla
		\item Height_Stricta
		\item m
		\item Center_x
		\item Center_y
		\item Rel_x
		\item Rel_y
}}
\format{A data frame with 575 rows and 26 variables}
\description{
Information about plots surveyed within the Guana Tolomato Matanzas National Estuarine Research Reserve
Survey information for plots containing cacti, surveyed during Summer 2012.
\itemize{
		\item NetworkID Do not use; number of unique Network ID's equals 86, not 88
		\item Island Identifying number/name for habitat patch
		\item unitID The Tag_Number of the plot
		\item n.surveys
		\item Date
		\item Network.mod Tag_Number identifier for each unique Network (n = 88)
		\item x Easting
		\item y Northing
		\item Stricta
		\item Pusilla
		\item Cactus
		\item CACA_on_Pusilla
		\item CACA_on_Stricta
		\item MEPR_on_Pusilla
		\item MEPR_on_Stricta
		\item Old_Moth_Evidence_Pusilla
		\item Old_Moth_Evidence_Stricta
		\item Percent_Cover_Pusilla
		\item Percent_Cover_Stricta
		\item Height_Pusilla
		\item Height_Stricta
		\item m
		\item Center_x
		\item Center_y
		\item Rel_x
		\item Rel_y
}
}

