% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_var.R
\name{pop_var}
\alias{pop_var}
\title{Population Variance}
\usage{
pop_var(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Population variance
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
