% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopulationSummaryStatistics.R
\name{popCV}
\alias{popCV}
\title{Population Coefficient of Variation}
\usage{
popCV(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Population Coefficient of Variation
}
