% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopulationSummaryStatistics.R
\name{PopVariance}
\alias{PopVariance}
\title{Return the population variance of a vector, after removing NAs, and round.}
\usage{
PopVariance(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}
