% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hat.R
\name{R_hat}
\alias{R_hat}
\title{Calculate the Horvitz-Thompson Ratio Estimator}
\usage{
R_hat(y, x, N, n1, m_vec, replace = "FALSE")
}
\arguments{
\item{y}{The variable of interest, $y$. Must be a numeric vector. The criterion that determines whether adaptive cluster sampling takes place is based on this variable.}

\item{x}{A vector of values for variable $x$, which is auxiliary data about the variable of interest $y$.}

\item{replace}{Whether sampling should be done with or without replacement. Defaults to \code{FALSE}.}
}
\description{
R_hat is calculated by dividing the Horvitz-Thompson estimator, Tau_hat_z by Tau_hat_x. See p. 77, Thompson (2002). THERE CANNOT BE NAS IN X OR Y?
}
\examples{
@template ex_Thompson2002_p_78
}
