% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations.R
\name{sampleRealizations}
\alias{sampleRealizations}
\title{Sample species patch realizations simulations}
\usage{
sampleRealizations(popdata, sims, n1, avar = NULL, ovar, rvar = NULL,
  SamplingDesign = "ACS", yvar, y_HT_formula = "y_HT",
  var_formula = "var_y_HT", mThreshold = NULL, f_max = 2,
  SampleEstimators = FALSE, SpatStat = TRUE, mChar = TRUE,
  popvar = "n.networks", realvar = "realization", weights = "S")
}
\arguments{
\item{popdata}{patch realizations}

\item{sims}{Number of simulations per population.}

\item{n1}{Vector of initial sample size(s) for the initial simple random sample(s) without replacement; can be a single value or vector of values}

\item{avar}{Vector of variables for which abundance should be estimated.}

\item{ovar}{Vector of variables for which occupancy should be estimated.}

\item{rvar}{Vector of variables for which ratio estimators should be used.}

\item{SamplingDesign}{Whether restricted or unrestricted adaptive cluster sampling should be performed; defaults to \code{FALSE}.}

\item{yvar}{variable upon which adaptive cluster sampling criterion is based}

\item{y_HT_formula}{Default is "Thompson".}

\item{var_formula}{Default is "var_y_HT".}

\item{f_max}{Default is 2.}

\item{SampleEstimators}{If "TRUE", calculate the sample mean and sample variance for each simulation. Default is FALSE.}

\item{SpatStat}{If "TRUE", for each simulation calculate Moran's I, and the nugget, sill, and range of the semivariogram. Default is TRUE}

\item{mChar}{If "TRUE", for each simulation calculate summary statistics (median, mean, min, and max) for the sample's m values. Also, for each simulation and for the set of unique m values, calculate the same summary statistics.}

\item{popvar}{Default is "n.networks"}

\item{realvar}{Default is "realization"}

\item{weights}{If SpatStat is "TRUE", this is a vector giving spatial weight matrix styles to use to calculate the Join Count and Moran's I statistics. Can take on values "W", "B", "C", "U", "S", and "minmax". See nb2listw for more details.}

\item{mThrshld}{Default is NULL.}
}
\description{
This function simulates sampling of multiple realizations of patches of the species of interest within the grid of locations created with \code{createPopulation}.
}
\examples{
sims=200
n1=c(5,10,20,40)
population <- createPop(x_start = 1, x_end = 30, y_start = 1, y_end = 30)
avar = NULL
ovar = c(
	"Stricta",
	"Pusilla",
"Cactus",
"CACA_on_Pusilla",
"CACA_on_Stricta",
"MEPR_on_Pusilla",
"MEPR_on_Stricta",
"Old_Moth_Evidence_Pusilla",
"Old_Moth_Evidence_Stricta"
# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
# "Percent_Cover_Stricta",
# "Height_Pusilla",
# "Height_Stricta",
)		
data(CactusRealizations)
popdata = CactusRealizations
simulation_data <- sampleRealizations(
popdata, 
sims, 
n1, 
avar, 
ovar, 
popvar="population", 
yvar="Cactus"
)
# sims=200
# #n1=c(75,150,225,300,350)
simulation_data_SRSWOR <- sampleRealizations(
	popdata, 
	sims, 
	n1, 
	avar, 
	ovar,
	popvar="population"
)
# # save data
# write.csv(simulation_data_SRSWOR, file=paste("simulation_data_SRSWOR", 
# format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M"), ".csv", sep=""))
#  write.csv(simulation_data, file=paste("simulation_data", format(Sys.time(), 
# "\%Y-\%m-\%d_\%H-\%M"), ".csv", sep=""))
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
