% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleSpeciesPatchRealizations.R
\name{sampleSpeciesPatchRealizations}
\alias{sampleSpeciesPatchRealizations}
\title{cactus patch realizations simulations}
\usage{
sampleSpeciesPatchRealizations(patchdat, simulations, nsamples, avar, ovar,
  rvar, SamplingDesign = "ACS", y_variable, y_HT_formula = "y_HT",
  var_y_HT_formula = "var_y_HT", m_threshold = NULL, f_max = 2,
  SampleEstimators = FALSE, SpatialStatistics = TRUE,
  mCharacteristics = TRUE, patch_variable = "n.networks",
  realization_variable = "realization", weights = "S")
}
\arguments{
\item{patchdat}{patch realizations}

\item{simulations}{Number of simulations per population.}

\item{nsamples}{Vector of initial sample size(s) for the initial simple random sample(s) without replacement; can be a single value or vector of values}

\item{avar}{Vector of variables for which abundance should be estimated.}

\item{ovar}{Vector of variables for which occupancy should be estimated.}

\item{rvar}{Vector of variables for which ratio estimators should be used.}

\item{SamplingDesign}{Whether restricted or unrestricted adaptive cluster sampling should be performed; defaults to \code{FALSE}.}

\item{y_variable}{variable upon which adaptive cluster sampling criterion is based}

\item{y_HT_formula}{Default is "Thompson".}

\item{m_threshold}{Default is NULL.}

\item{f_max}{Default is 2.}

\item{SampleEstimators}{If "TRUE", calculate the sample mean and sample variance for each simulation. Default is FALSE.}

\item{SpatialStatistics}{If "TRUE", for each simulation calculate Moran's I, and the nugget, sill, and range of the semivariogram. Default is TRUE}

\item{mCharacteristics}{If "TRUE", for each simulation calculate summary statistics (median, mean, min, and max) for the sample's m values. Also, for each simulation and for the set of unique m values, calculate the same summary statistics.}

\item{patch_variable}{Default is "n.networks"}

\item{realization_variable}{Default is "realization"}

\item{weights}{If SpatialStatistics is "TRUE", this is a vector giving spatial weight matrix styles to use to calculate the Join Count and Moran's I statistics. Can take on values "W", "B", "C", "U", "S", and "minmax". See nb2listw for more details.}
}
\description{
This function simulates sampling of multiple realizations of patches of the species of interest within the grid of locations created with \code{createPopulation}.
}
\examples{
simulations=200
nsamples=c(5,10,20,40)
population <- createPopulation(x_start = 1, x_end = 30, y_start = 1, y_end = 30)
avar = NULL
ovar = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)		
# patchdat = cactus.realizations
# simulation_data <- sampleSpeciesPatchRealizations(patchdat, simulations, 
# 	nsamples, population, avar, ovar)
# simulations=200
# #nsamples=c(75,150,225,300,350)
# simulation_data_SRSWOR <- sampleSpeciesPatchRealizations(patchdat, 
# 	simulations, nsamples, population, avar, ovar)
# # save data
# write.csv(simulation_data_SRSWOR, file=paste("simulation_data_SRSWOR", 
# format(Sys.time(), "\%Y-\%m-\%d_\%H-\%M"), ".csv", sep=""))
#  write.csv(simulation_data, file=paste("simulation_data", format(Sys.time(), 
# "\%Y-\%m-\%d_\%H-\%M"), ".csv", sep=""))
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
