% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopulationSummaryStatistics.R
\name{Sum}
\alias{Sum}
\title{Return the sum of a vector, after removing NAs.}
\usage{
Sum(x)
}
\arguments{
\item{x}{Vectors of data.}
}
\description{
Written for use in the tables::tabular function to create publication-ready tables.
}
