% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeNetworkInformation.R
\name{summarizeNetworkInformation}
\alias{summarizeNetworkInformation}
\title{Calculate the number of units per network and the unit inclusion probability for each network.}
\usage{
summarizeNetworkInformation(patch_data, variables, grouping.variables = NULL,
  nsamples, m_var)
}
\arguments{
\item{patch_data}{patch realization data}

\item{variables}{variables to summarise}

\item{grouping.variables}{categories that group networks}

\item{nsamples}{Size(s) of the initial simple random sample(s) without replacement. Can be a single value or a vector values.}

\item{m_var}{The variable on which the ACS sampling criterion is based}
}
\value{
vector of inclusion probabilities
}
\description{
The purpose of this is to reduce computation time by calculating some necessary information before the data is boostrapped. The function calculates (1) $pi_i$ (the unit inclusion probability) for each network and sample size (\code{nsamples}) and (2) the number of units per network for each of the \code{variables}.
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}

