% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thompson1990Fig1Sample.R,
%   R/Thompson1990Figure1Sample.R
\name{Thompson1990Fig1Sample}
\alias{Thompson1990Fig1Sample}
\title{Create the population displayed in Figure 1 from Thompson (1990)}
\format{A data frame with 10 rows and 2 variables}
\usage{
Thompson1990Fig1Sample()
}
\value{
The population displayed in Figure 1 from Thompson (1990).
}
\description{
An example dataset from Figure 1 from Thompson (1990) containing the x- and y-coordinates of units selected in a simple random sample without replacement.

\itemize{
  \item x. X coordinate.
  \item y. Y coordinate.
}
}
\examples{
createThompson1990Fig1Sample()
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}
