% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thompson1990Figure1Sample.R
\name{Thompson1990Figure1Sample}
\alias{Thompson1990Figure1Sample}
\title{Sample from Figure 1 from Thompson (1990)}
\format{A data frame with 10 rows and 2 variables}
\description{
An example dataset from Figure 1 from Thompson (1990) containing the locations of the initial sample.
}
\details{
\itemize{
  \item x. X coordinate.
  \item y. Y coordinate.
}
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

