% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thompson1990Table1data.R
\name{Thompson1990Table1data}
\alias{Thompson1990Table1data}
\title{Table 1 data from Thompson (1990)}
\format{A data frame with 10 rows and 2 variables}
\description{
Data from Table 1 from Thompson (1990).
}
\details{
\itemize{
  \item x. X coordinate.
  \item y. Y coordinate.
}
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

