% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{var_pi}
\alias{var_pi}
\title{Variance estimator free of joint inclusion probability calculations for unequal probability sampling}
\usage{
var_pi(n, y, pi_i_values, estimator)
}
\arguments{
\item{y}{Vector of $y$ values.}

\item{pi_i_values}{vector of first-order inclusion probabilities, calculated using \code{pi_i}.}

\item{estimator}{Options include "Hartley_Rao", "Hajek", "Rosen", "Berger", and "Deville".}

\item{N}{Population size}
}
\description{
Variance estimator free of joint inclusion probability calculations for unequal probability sampling
}
\references{
Hajek, J. (1964). Asymptotic theory of rejective sampling with varying probabilities from a finite population. The Annals of Mathematical Statistics.
Berger, Y. G., & Tille, Y. (2009). Sampling with Unequal Probabilities. Handbook of Statistics, 1–17.
}

