% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_R_hat_C.R
\name{var_R_hat}
\alias{var_R_hat}
\title{Calculate the variance of R hat}
\usage{
var_R_hat(y, x, N, n1, m_vec, replace = "FALSE")
}
\arguments{
\item{y}{The variable of interest, \eqn{y}. Must be a numeric vector. The criterion that determines whether adaptive cluster sampling takes place is based on this variable.}

\item{x}{A vector of values for variable $x$, which is auxiliary data about the variable of interest $y$.}

\item{N}{Population size.}

\item{n1}{An integer giving the initial sample size (ie., the size of the initial simple random sample).}

\item{m_vec}{Vector of values \eqn{m} for the set of units in a sample, of length \eqn{n1}. Each \eqn{m} value within the vector \code{m_vec} denotes the number of units satisfying the ACS criterion for the network \eqn{i} to which the unit belongs.}

\item{replace}{Whether sampling should be done with or without replacement. Defaults to \code{FALSE}.}
}
\description{
Calculate the variance of R hat

# Example from Thompson (2002), p. 78-79
N = 100
n1 = 4
x = c(60, 60, 14, 1)
y = c(1, 1, 1, 1)
m = c(5, 5, 2, 1)
R_hat(y, x, N, n1, m, replace="TRUE")
var_R_hat(y, x, N, n1, m)
}
