% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{var_Tille}
\alias{var_Tille}
\title{Tille (2006) variance estimator free of joint inclusion probability calculations for unequal probability sampling}
\usage{
var_Tille(n, y, pi_i_values)
}
\arguments{
\item{n}{sample size}

\item{y}{Vector of $y$ values.}

\item{pi_i_values}{vector of first-order inclusion probabilities, calculated using \code{Hajek}.
Tille, Y. (2006).}
}
\description{
Tille (2006) variance estimator free of joint inclusion probability calculations for unequal probability sampling
}
