% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_x_HT_C.R
\name{var_x_HT}
\alias{var_x_HT}
\title{Calculate the variance of the Horvitz-Thompson estimator of the mean}
\usage{
var_x_HT(N, n1, m, x, pi_i_values = NULL)
}
\arguments{
\item{N}{Population size}

\item{n1}{Initial sample size}

\item{m}{Vector of $m$, each corresponding to a unique network.}

\item{x}{Vector of $x$ total, each corresponding to a unique network.}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}
}
\description{
Calculate the variance of the Horvitz-Thompson estimator of the mean
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)
# Sampling of population from Figure 1, Thompson (1990)

data(Thompson1990Figure1Population)
data(Thompson1990Figure1Sample)

# plot sample overlaid onto population
ggplot() +
	geom_point(data=Thompson1990Figure1Population, aes(x,y, 
	size=factor(x_value),
		shape=factor(x_value))) +
	scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	geom_point(data=Thompson1990Figure1Sample, aes(x,y), shape=0, size=7)

# INITIATE ACS
Z = createACS(population=Thompson1990Figure1Population, 
	n1=dim(Thompson1990Figure1Sample)[1], 
	initial_sample=Thompson1990Figure1Sample, y_variable="x_value")

# CALCULATE var(x_HT)
# create dataframe of network info
Z_summary <- Z \%>\% group_by(NetworkID) \%>\%
	summarise(
		m = m[1],
		x_total = sum(x_value, rm.na=TRUE)
		) \%>\%
		filter(NetworkID > 0)

var_x_HT(
	N = dim(Thompson1990Figure1Population)[1], 
	n1 = dim(Thompson1990Figure1Sample)[1], 
	m = Z_summary$m, 
	x = Z_summary$x_total
)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

