% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationResults.R
\name{xscale}
\alias{xscale}
\title{Function to Calculate Numbers to Label X-axis}
\usage{
xscale(dig = 2, ...)
}
\arguments{
\item{x}{vector of data}
}
\description{
Function to Calculate Numbers to Label X-axis
}
