% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNetworkMembership.R
\name{assignNetworkMembership}
\alias{assignNetworkMembership}
\title{Assign network membership to adaptive cluster sample units.}
\usage{
assignNetworkMembership(popdata, plot.size = 1)
}
\arguments{
\item{popdata}{A data frame providing information about the population to be sampled. Required columns include geographic coordinate columns \code{x} and \code{y}, as well as a column containing the variable of interest, whose name is provided to the function via the \code{yvar} argument.}

\item{plot.size}{The length and width of plots, in coordinate units. Defaults to 1.}
}
\value{
Returns the popdata data.frame with a new column for Network ID as well as a column for m, the number of units in each network.
}
\description{
This function assigns network membership to units in an adaptive cluster sample; if units are neighbors, they are assigned the same network ID.
}
\examples{
data(Thompson1990Fig1Pop)

# plot species abundance
library(magrittr)
library(ggplot2)
library(dplyr)
dat <- Thompson1990Fig1Pop \%>\% 
	filter(y_value > 0)
ggplot(dat, aes(x,y, size=y_value)) + geom_point()

# assign network membership of units containing the species of interest
P_networks <- assignNetworkMembership(Thompson1990Fig1Pop \%>\%
	 filter(y_value > 0))

# plot networks
ggplot(
	data=P_networks, 
	aes(x,y, size=y_value, shape=factor(NetworkID))
) + geom_point()
# coordinates should be given as x and y
}
\references{
\insertRef{saubyadaptive}{ACSampling}
}
