% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assignNetworkMembership.R
\name{assignNetworkMembership}
\alias{assignNetworkMembership}
\title{Assign network membership to adaptive cluster sample units.}
\usage{
assignNetworkMembership(popdata, plot.size = 1)
}
\arguments{
\item{popdata}{Dataset to be used, containing information about units containing the species of interest, including x and y coordinates, with column names "x" and "y," respectively.}

\item{plot.size}{The length and width of plots, in coordinate units. Defaults to 1.}
}
\value{
Returns popdata with a new column for Network ID as well as a column for m, the number of units in the network.
}
\description{
This function assigns network membership to units in an adaptive cluster sample; if units are neighbors, they are assigned to the same network ID.
}
\examples{
data(Thompson1990Figure1Population)

# plot species abundance
library(magrittr)
library(ggplot2)
library(dplyr)
ggplot(data=Thompson1990Fig1Pop \%>\% 
	filter(y_value > 0), aes(x,y, size=factor(y_value))) + geom_point()

# assign network membership of units containing the species of interest
P_networks <- assignNetworkMembership(Thompson1990Fig1Pop \%>\%
	 filter(y_value > 0))

# plot networks
ggplot(
	data=P_networks, 
	aes(x,y, size=factor(y_value), shape=factor(NetworkID))
) + geom_point()
# coordinates should be given as x and y
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
