% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePopSummaryStats.R
\name{calcPopSummaryStats}
\alias{calcPopSummaryStats}
\title{summaryvar = ovar
# WHAT WAS I THINK HERE? for grouping variables?
popgroupvar = "n.networks" # c("n.networks", "realization")
# create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
start.seed=1
buffer=5
rvar = c("CACA_on_Stricta", "CACA_on_Pusilla")
cactus.realizations <- createRealizations(x_start, x_end,
	y_start, y_end, buffer, n.networks, n.realizations, SpeciesInfo, start.seed,
	ovar, "Cactus")
patch_data_summary <- calcPopSummaryStats(popdata=cactus.realizations,
	summaryvar=ovar, popgroupvar=popgroupvar, nrow=30, ncol=30)}
\usage{
calcPopSummaryStats(
  popdata,
  summaryvar,
  rvar = NULL,
  popgroupvar,
  spatweights = "S",
  nrow,
  ncol
)
}
\description{
summaryvar = ovar
# WHAT WAS I THINK HERE? for grouping variables?
popgroupvar = "n.networks" # c("n.networks", "realization")
# create realizations
x_start = 1
x_end = 30
y_start = 1
y_end = 30
n.networks = c(5, 15, 10, 20, 30, 40)
n.realizations = 1
SpeciesInfo = PlotSurveys_season1
start.seed=1
buffer=5
rvar = c("CACA_on_Stricta", "CACA_on_Pusilla")
cactus.realizations <- createRealizations(x_start, x_end,
	y_start, y_end, buffer, n.networks, n.realizations, SpeciesInfo, start.seed,
	ovar, "Cactus")
patch_data_summary <- calcPopSummaryStats(popdata=cactus.realizations,
	summaryvar=ovar, popgroupvar=popgroupvar, nrow=30, ncol=30)
}
