% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations_depends_estimators.R
\name{calcSpatStats}
\alias{calcSpatStats}
\title{Calculate Spatial Statistics}
\usage{
calcSpatStats(alldata_all, weights)
}
\arguments{
\item{alldata_all}{NEED TO DESCRIBE.}

\item{weights}{String identifying the weight to use. See REFERENCE FOR?}
}
\description{
Calculate Spatial Statistics
}
