% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSSQR.R
\name{calcSSQR}
\alias{calcSSQR}
\title{Calculate SSQ_R}
\usage{
calcSSQR(popdata, variable, popvar = NULL)
}
\arguments{
\item{popdata}{Information about the populations of interest. MORE DETAIL ABOUT STRUCTURE OF THIS.}

\item{variable}{Variable for which to calculate SSQr.}

\item{popvar}{A character string (OR VECTOR??) used to identify different populations.
The default value is NULL.}
}
\value{
Dataframe including original data and RE estimates.
}
\description{
SSQ_R is the ratio of the within-network sum of squares to the total sum of squares. The adaptive cluster sampling (ACS) design becomes more efficient relative to simple random sampling (SRS) as the within-network variation increases relative to overall variation [@thompson1996adaptive]. We calculate \deqn{SSQ_R = SSQ_w/SSQ_\tau}{SSQ_R = SSQ_w/SSQ_tau} as the ratio of the within-network sum of squares \deqn{SSQ_w = \sum_{j=1}^{\kappa} \sum_{i \in j} (y_{j,i} - \bar{y_{j}})^2}{SSQ_w = \sum_{j=1}^{\kappa} \sum_{i \in j} (y_{j,i} - \bar{y_{j}})^2}, to the total sum of squares \deqn{SSQ_\tau= \sum_{i=1}^{N} (y_i - \mu)^2}{SSQ_\tau= \sum_{i=1}^{N} (y_i - \mu)^2}. Thus, an increase in \code{SSQ_R} indicates an increase in the efficiency of the ACS design relative to SRS.
}
\examples{
library(magrittr)
library(dplyr)
data(Thompson1990Fig1Pop)
temp <- Thompson1990Fig1Pop \%>\%
	mutate(pop = 1)
popdata <- temp
variable <- "y_value"
popvar <- "pop"
# calcSSQR(popdata, variable, popvar)
popvar <- NA
# calcSSQR(popdata, variable, popvar)
}
\references{
\insertRef{su2003estimator}{ACSampling}
}
