% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateJointInclusionProbabilities.R
\name{calculateJointInclusionProbabilities}
\alias{calculateJointInclusionProbabilities}
\title{Calculate Joint Inclusion Probabilities Using Simulations}
\usage{
calculateJointInclusionProbabilities(patchdat = patch_data_3, simulations,
  nsamples, SamplingDesign = "ACS", y_variable, f_max = NULL,
  population.grouping.variable = NULL)
}
\arguments{
\item{patchdat}{patch realizations}

\item{simulations}{Number of simulations per population.}

\item{nsamples}{Vector of initial sample size(s) for the initial simple random sample(s) without replacement; can be a single value or vector of values}

\item{SamplingDesign}{Sampling design; ACS or RACS.}

\item{y_variable}{variable upon which adaptive cluster sampling criterion is based
population.grouping.variable variable identifying unique populations}
}
\description{
Calculate inclusion probabilities for each unit in a population using simulations.
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}

