% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateSamplingBias.R
\name{calculateRE}
\alias{calculateRE}
\title{Calculate Relative Efficiency (RE)}
\usage{
calculateRE(MSE_ComparisonSamplingDesign = MSE_ComparisonSamplingDesign,
  MSE_BaselineSamplingDesign = MSE_SRSWOR,
  grouping.variables = grouping.variables)
}
\arguments{
\item{MSE_ComparisonSamplingDesign}{Sampling design for which relative efficiency (RE) should be calculated.}

\item{MSE_BaselineSamplingDesign}{The sampling design to which the "comparison sampling design" is compared to and its efficiency, relative to this one, is calculated.}

\item{grouping.variables}{Categorical variables identifying the patch realization from which the simulation data was generated (e.g., \code{n.networks} and \code{realization}).}
}
\value{
Dataframe including original data and RE estimates.
}
\description{
Calculate Relative Efficiency (RE)
}

