% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createACS.R
\name{createACS}
\alias{createACS}
\title{Create an Adaptive Cluster Sample.}
\usage{
createACS(popdata, n1, yvar, criterion = 0, seed = NA, initsample = NA)
}
\arguments{
\item{popdata}{A data frame providing information about the population to be sampled. Required columns include geographic coordinate columns \code{x} and \code{y}, as well as a column containing the variable of interest, whose name is provided to the function via the \code{yvar} argument.}

\item{n1}{An integer giving the initial sample size (ie., the size of the initial simple random sample).}

\item{yvar}{The variable of interest, $y$, that determines the condition under which adaptive cluster sampling takes place. Must be a numeric value.}

\item{criterion}{Numeric threshold value of \code{yvar} that initiates ACS. Defaults to 0 (ie., anything greater than 0 initiates adaptive cluster sampling).}

\item{seed}{Optional. A vector of integers to feed to \code{set.seed()} so that the sampling is reproducible.}

\item{initsample}{Optional. A data frame that includes geographic coordinate columns \code{x} and \code{y} of an the initial sample. The defaults is \code{NA}, which results in the initial sample being selected according to simple random sampling without replacement.}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Create an Adaptive Cluster Sample.
}
\examples{
library(ggplot2)
data(Thompson1990Fig1Pop)
data(Thompson1990Figure1Sample)

# Initiate ACS
Z = createACS(
   popdata=Thompson1990Fig1Pop, 
   seed=9, 
   n1=10, 
   yvar="y_value", 
   criterion=0
)

# plot ACS sample overlaid onto population
ggplot() +
   geom_point(
	 data=Thompson1990Fig1Pop, 
	 aes(x, y, size=factor(y_value),shape=factor(y_value))
) +
	 scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	 geom_point(data=Z, aes(x,y), shape=0, size=7)
# Initiate ACS, different seed
Z = createACS(
   popdata=Thompson1990Fig1Pop, 
   seed=26, 
   n1=10, 
   yvar="y_value", 
  criterion=0
)

# plot ACS sample overlaid onto population
ggplot() +
  geom_point(
     data=Thompson1990Fig1Pop, 
     aes(x, y, size=factor(y_value), shape=factor(y_value))
  ) +
	scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	geom_point(data=Z, aes(x,y), shape=0, size=7)
}
\references{
@template Thompson1990
}
