% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createACS.R
\name{createACS}
\alias{createACS}
\title{Create an Adaptive Cluster Sample.}
\usage{
createACS(population, seed = 1, n1, y_variable, condition = 0,
  initial_sample = NA)
}
\arguments{
\item{population}{grid of population to be sampled.}

\item{seed}{vector of numbers to feed to \code{set.seed()} so that the sampling is reproducible.}

\item{n1}{initial sample size (sampled according to simple random sampling without replacement).}

\item{y_variable}{Variable of interest, used to determine condition under which adaptive cluster sampling takes place.}

\item{condition}{Threshold value of the y variable that initiates ACS. Defaults to \code{0}.}

\item{initial_sample}{List of x and y coordinates of the initial sample. Defaults to "NA" so that the initial sample is selected according to simple random sampling without replacement.}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Create an Adaptive Cluster Sample.
}
\examples{
library(ggplot2)
data(Thompson1990Figure1Population)
data(Thompson1990Figure1Sample)

# Initiate ACS
Z = createACS(Thompson1990Figure1Population, seed=2, n1=10, "y_value", condition=0)

# plot ACS sample overlaid onto population
ggplot() +
	geom_point(data=Thompson1990Figure1Population, aes(x,y, size=factor(y_value),
		shape=factor(y_value))) +
	scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	geom_point(data=Z, aes(x,y), shape=0, size=7)
# Initiate ACS, different seed
Z = createACS(Thompson1990Figure1Population, seed=26, n1=10, "y_value", condition=0)

# plot ACS sample overlaid onto population
ggplot() +
	geom_point(data=Thompson1990Figure1Population, aes(x,y, size=factor(y_value),
		shape=factor(y_value))) +
	scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	geom_point(data=Z, aes(x,y), shape=0, size=7)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} A Sampling Strategy Designed to Maximize the Efficiency of Data Collection of Food Web Relationships.
}

