% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCactusRealizations.R
\name{createCactusRealizations}
\alias{createCactusRealizations}
\title{Create Cactus Realization Data}
\usage{
createCactusRealizations(PlotSurveys_season1, ovar)
}
\arguments{
\item{PlotSurveys_season1}{Plot survey data from season 1.}

\item{ovar}{occupancy variables.}
}
\value{
The data from Sauby and Christman.
}
\description{
Create Cactus Realization Data
}
\examples{
library(ggplot2)
CactusRealizations <- createCactusRealizations(
	PlotSurveys_season1, 
	ovar = c(
	"Stricta",
	"Pusilla",
	"Cactus",
	"CACA_on_Pusilla",
	"CACA_on_Stricta",
	"MEPR_on_Pusilla",
	"MEPR_on_Stricta",
	"Old_Moth_Evidence_Pusilla",
	"Old_Moth_Evidence_Stricta"
	# "Percent_Cover_Pusilla", # how do I do these? they are occupancy nor abundance
	# "Percent_Cover_Stricta",
	# "Height_Pusilla",
	# "Height_Stricta",
)
)
ggplot(
	CactusRealizations,
	aes(
		x, 
		y, 
		shape = factor(Cactus)
)) +
	geom_point() +
	facet_wrap(~population) +
	scale_shape_manual(values=c(4,16))
}
\references{
\insertRef{saubyadaptive}{ACSampling}
}
