% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCactusRealizationSummary.R
\name{createCactusRealizationSummary}
\alias{createCactusRealizationSummary}
\title{Create Summary of Cactus Realization Data}
\usage{
createCactusRealizationSummary(CactusRealizations)
}
\arguments{
\item{CactusRealizations}{the data to give the function}
}
\value{
The data from Sauby and Christman.
}
\description{
Create Summary of Cactus Realization Data
}
\references{
\insertRef{saubyadaptive}{ACSampling}
}
