% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createlambdap_10_tau_25.R
\name{createlambdap_10_tau_25}
\alias{createlambdap_10_tau_25}
\title{Create the population, $\lambda_P$ = 5, $tau$ = 10, from Christman (1997)}
\usage{
createlambdap_10_tau_25()
}
\value{
The population displayed in Christman (1997), where $\lambda_P$ = 5, $tau$ = 10.
}
\description{
Create the population, $\lambda_P$ = 5, $tau$ = 10, from Christman (1997)
}
\examples{
library(ggplot2)
data(lambdap_10_tau_25)

ggplot(lambdap_10_tau_25 \%>\% filter(y_value!=0), aes(x, y)) + geom_text(aes(label=y_value))
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
Christman, M. C. (1997). Efficiency of some sampling designs for spatially clustered populations. \emph{Environmetrics}, 8: 145--166.
}

