% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRACS_flex.R
\name{createRACS}
\alias{createRACS}
\title{Create a Restricted Adaptive Cluster Sample}
\usage{
createRACS(
  popdata,
  n1,
  yvar,
  criterion = 0,
  seed = NA,
  initsample = NULL,
  f_max = 2
)
}
\arguments{
\item{popdata}{A data frame providing information about the population to be sampled. Required columns include geographic coordinate columns \code{x} and \code{y}, as well as a column containing the variable of interest, whose name is provided to the function via the \code{yvar} argument.}

\item{n1}{An integer giving the initial sample size (e.g., a simple random sample).}

\item{yvar}{A string giving the name of the variable of interest, \eqn{y}, in the supplied dataframe \eqn{popdata}. This variable determines the condition under which adaptive cluster sampling takes place. In the dataframe $popdata$ this variable \eqn{y} must be numeric.}

\item{criterion}{Numeric threshold value of the variable of interest \eqn{y} (whose name in the dataframe $popdata$ is supplied via the \code{yvar} argument) that initiates ACS. Defaults to 0 (ie., anything greater than 0 initiates adaptive cluster sampling).}

\item{seed}{Optional. An integer to feed to \code{set.seed()} so that the sampling is reproducible.}

\item{initsample}{Optional. A data frame that includes geographic coordinate columns \code{x} and \code{y} of the initial sample. The defaults is \code{NA}, which results in the initial sample being selected according to simple random sampling without replacement.}

\item{f_max}{The maximum number of expanding sets of adjacent units surveyed
per primary unit in the adaptive phase of ACS. For more information, see 
Sauby and Christman (in prep).}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Select a restricted adaptive cluster sample from a dataframe containing columns indicating the spatial coordinates of each unit (\code{x} and \code{y}) as well as the variable of interest, provided to the function via the \code{yvar} argument. Cluster sampling is initiated for primary units in which the value of the variable of interest exceeds the criterion value.
The criterion value can be set with the \code{criterion} argument, which defaults to zero. Cluster sampling is restricted by the argument, \code{f_max}.
}
\examples{
library(ggplot2)
seed=3
n1=5
y_variable = "y_value"
f_max = 3
# cluster sampling not triggered, only y_values of zero sampled
Z = createRACS(
	popdata=lambdap_5_tau_1, 
	n1=n1, 
	yvar=y_variable, 
	seed=seed, 
	f_max=f_max
)

ggplot() +
   geom_point(
      data=lambdap_5_tau_1,
      aes(x,y, size=y_value)
   ) +
   geom_point(data=Z, aes(x,y), shape = 0, size=7) 
}
\references{
\insertRef{saubyadaptive}{ACSampling}
}
