% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRACS_flex.R
\name{createRACS}
\alias{createRACS}
\title{Create a Restricted Adaptive Cluster Sample, for any step size}
\usage{
createRACS(
  popdata,
  n1,
  yvar,
  criterion = 0,
  seed = NA,
  initsample = NULL,
  f_max = 2
)
}
\arguments{
\item{popdata}{A data frame providing information about the population to be sampled. Required columns include geographic coordinate columns \code{x} and \code{y}, as well as a column containing the variable of interest, whose name is provided to the function via the \code{yvar} argument.}

\item{n1}{An integer giving the initial sample size (ie., the size of the initial simple random sample).}

\item{yvar}{The variable of interest, $y$, that determines the condition under which adaptive cluster sampling takes place. Must be a numeric value.}

\item{criterion}{Numeric threshold value of \code{yvar} that initiates ACS. Defaults to 0 (ie., anything greater than 0 initiates adaptive cluster sampling).}

\item{seed}{Optional. A vector of integers to feed to \code{set.seed()} so that the sampling is reproducible.}

\item{initsample}{Optional. A data frame that includes geographic coordinate columns \code{x} and \code{y} of an the initial sample. The defaults is \code{NA}, which results in the initial sample being selected according to simple random sampling without replacement.}

\item{f_max}{WHAT IS IT}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Create a Restricted Adaptive Cluster Sample, for any step size

Sauby, K.E and Christman, M.C. \emph{In preparation.} A Sampling Strategy Designed to Maximize the Efficiency of Data Collection of Food Web Relationships
}
\examples{
library(ggplot2)
popdata = lambdap_5_tau_1
seed=3
n1=5
yvar = "y_value"
f_max = 3
Z = createRACS(
	popdata = lambdap_5_tau_1, 
	n1 = n1, 
	yvar = yvar, 
	seed = seed, 
	f_max = f_max
)
ggplot() +
geom_point(
data=Z, 
aes(x,y, size=factor(y_value), shape=factor(y_value))
) +
scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
geom_point(data=Z, aes(x,y), shape = 0, size=7) +
ggtitle("f_max = 1")
}
