% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createRACS.R
\name{createRACS}
\alias{createRACS}
\title{Create a Restricted Adaptive Cluster Sample}
\usage{
createRACS(population, n1, y_variable, condition = 0, seed = NA,
  initial_sample = NA)
}
\arguments{
\item{population}{grid of population to be sampled.}

\item{n1}{initial sample size (sampled according to simple random sampling without replacement).}

\item{y_variable}{Variable of interest, used to determine condition under which adaptive cluster sampling takes place.}

\item{condition}{Threshold value of the y variable that initiates Restricted ACS. Defaults to \code{0}.}

\item{seed}{vector of numbers to feed to \code{set.seed()} so that the sampling is reproducible.}

\item{initial_sample}{List of x and y coordinates of the initial sample. Defaults to "NA" so that the initial sample is selected according to simple random sampling without replacement.}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Create a Restricted Adaptive Cluster Sample
}
\examples{
library(ggplot2)
Z = createRACS(Thompson1990Figure1Population, seed=26, n1=10, "y_value")
ggplot() +
	geom_point(data=Thompson1990Figure1Population, aes(x,y,shape=factor(y_value))) +
	scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	geom_text(data=Z, aes(x,y, label=NetworkID), shape=0, size=7)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} A Sampling Strategy Designed to Maximize the Efficiency of Data Collection of Food Web Relationships.
}
