% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSRS.R
\name{createSRS}
\alias{createSRS}
\title{Create a Simple Random Sample Without Replacement.}
\usage{
createSRS(popdata, n1, seed = NA, replace = F)
}
\arguments{
\item{popdata}{grid of population to be sampled.}

\item{n1}{initial sample size (sampled according to simple random sampling without replacement).}

\item{seed}{vector of numbers to feed to \code{set.seed()} so that the sampling is reproducible. Defaults to NA so that it is not necessary to specific a random number seed.}

\item{replace}{Should sampling be done with replacement? Defaults to FALSE.}
}
\value{
A restricted adaptive cluster sample.
}
\description{
Create a Simple Random Sample Without Replacement.
}
\examples{
# example
# create the population
# create the patch
# then sample

data(Thompson1990Figure1Population)
data(Thompson1990Figure1Sample)
Z = createSRS(Thompson1990Figure1Population, seed=2, n1=10)
}
