% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_pi_i.R
\name{estim_pi_i}
\alias{estim_pi_i}
\title{Estimate Inclusion Probabilities Using Simulations}
\usage{
estim_pi_i(
  popdata,
  sims,
  nsamples,
  SamplingDesign = "ACS",
  y_variable,
  f_max = NULL
)
}
\arguments{
\item{popdata}{Population data in the format XXXXXXXX}

\item{sims}{Number of simulations per population.}

\item{nsamples}{Vector of initial sample size(s) for the initial simple random sample(s) without replacement; can be a single value or vector of values}

\item{SamplingDesign}{Adaptive cluster sampling design (either "ACS" or "RACS").}

\item{y_variable}{Variable upon which adaptive cluster sampling criterion is based}

\item{f_max}{NEED DESCRIPTION HERE}
}
\value{
Returns a dataframe with the following columns:
realization
n.networks
N.SRSWOR.plots
SamplingDesign
simulations
coords
mean_m
max_m
min_m
median_m
}
\description{
Calculate inclusion probabilities for each unit in a realization using simulations.
}
\references{
\insertRef{saubyadaptive}{ACSampling}
}
