% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations.R
\name{getJoinCountTestEst}
\alias{getJoinCountTestEst}
\title{Calculate the Join Count Test estimate}
\usage{
getJoinCountTestEst(temp, lwb)
}
\arguments{
\item{temp}{dataset}

\item{lwb}{weights list}
}
\description{
Calculate the Join Count Test estimate
}
