% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations.R
\name{getMoranTestEst}
\alias{getMoranTestEst}
\title{Calculate the Moran Test estimate}
\usage{
getMoranTestEst(temp, lwb)
}
\arguments{
\item{temp}{dataset}

\item{lwb}{weights list}
}
\description{
Calculate the Moran Test estimate
}
