% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{Hajek}
\alias{Hajek}
\title{Hakey variance estimator free of joint inclusion probability calculations for unequal probability sampling}
\usage{
Hajek(pi_i, n)
}
\arguments{
\item{pi_i}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}

\item{n}{Initial sample size}
}
\description{
Option for the "var_pi" function.
}
\references{
Hajek, J. (1964). Asymptotic theory of rejective sampling with varying probabilities from a finite population. The Annals of Mathematical Statistics.
Berger, Y. G., & Tille, Y. (2009). Sampling with Unequal Probabilities. Handbook of Statistics, 1-17.
}
