% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{Hajek}
\alias{Hajek}
\title{Hakey variance estimator free of joint inclusion probability calculations for unequal probability sampling - or is N used instead of n. see p. 113 Tille new book}
\usage{
Hajek(pi_i, n)
}
\arguments{
\item{pi_i}{Need description here}

\item{n}{sample size (an integer value).}
}
\description{
Option for the "var_pi" function.
}
\references{
\insertRef{hajek1964asymptotic}{ACSampling}
\insertRef{berger2009sampling}{ACSampling}
}
