% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_HT.R
\name{new_y_HT}
\alias{new_y_HT}
\title{Calculate the Horvitz-Thompson mean of an adaptive cluster sample, NEW FORMULA.}
\usage{
new_y_HT(
  y,
  N,
  n1,
  m_threshold,
  pi_i_values = NULL,
  m_vec = NULL,
  sampling = NULL,
  criterion = NULL
)
}
\arguments{
\item{y}{The variable of interest, \eqn{y}. Must be a numeric vector. The criterion that determines whether adaptive cluster sampling takes place is based on this variable.}

\item{N}{Population size.}

\item{n1}{An integer giving the initial sample size (e.g., a simple random sample).}

\item{m_threshold}{threshold value above which to calculate \eqn{pi_i} and \eqn{pi_j} differently.}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}

\item{m_vec}{Vector of values \eqn{m} for the set of units in a sample, of length \eqn{n1}. Each \eqn{m} value within the vector \code{m_vec} denotes the number of units satisfying the ACS criterion for the network \eqn{i} to which the unit belongs.}

\item{sampling}{A vector (\code{character} format) describing whether units were included in the initial sample or subsequent ACS sample. Units selected in the initial sample should be given the value "Initial_Sample" in the \code{sampling} vector.}

\item{criterion}{Numeric threshold value of the variable of interest \eqn{y} (whose name in the dataframe $popdata$ is supplied via the \code{yvar} argument) that initiates ACS. Defaults to 0 (ie., anything greater than 0 initiates adaptive cluster sampling).}
}
\value{
The Horvitz-Thompson mean.
}
\description{
This calculate the Horvitz-Thompson mean of an adaptive cluster sample done by sampling without replacement.

where $v$ is the number of distinct units in the sample and
$J_k$ is an indicator variable, equalling 0 if the $k$ th unit in the sample does not satisfy the condition and was not selected in the initial sample; otherwise, $J_k = 1$.
}
\examples{
@template ex_Thompson2002_2_p_307_values
}
\references{
\insertRef{saubyadaptive}{ACSampling}
\insertRef{thompson1990adaptive}{ACSampling}
}
