% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_HT.R
\name{new_y_HT}
\alias{new_y_HT}
\title{Calculate the Horvitz-Thompson mean of an adaptive cluster sample, NEW FORMULA.}
\usage{
new_y_HT(y, N, n1, m_threshold, pi_i_values = NULL, m = NULL,
  sampling = NULL, criterion = NULL)
}
\arguments{
\item{y}{Attribute data about species of interest (e.g., abundance, presence/absence).}

\item{N}{Population size.}

\item{n1}{Initial sample size.}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}

\item{m}{Number of units satisfying the ACS criterion in network $i$.}

\item{sampling}{A vector (\code{character} format) describing whether units were included in the initial sample or subsequent ACS sample. Units selected in the initial sample should be given the value "Initial_Sample" in the \code{sampling} vector.}

\item{criterion}{The threshold value of \code{y} that triggers adaptive cluster sampling.}
}
\value{
The Horvitz-Thompson mean.
}
\description{
This calculate the Horvitz-Thompson mean of an adaptive cluster sample done by sampling without replacement.

where $v$ is the number of distinct units in the sample and
$J_k$ is an indicator variable, equalling 0 if the $k$ th unit in the sample does not satisfy the condition and was not selected in the initial sample; otherwise, $J_k = 1$.
}
\examples{
library(magrittr)
library(plyr)
library(dplyr)
library(ggplot2)

patch3 <- patch_data \%>\% filter(n.networks == levels(n.networks)[3])
newsample = createRACS(patch3, seed=26, n1=40, "Cactus")
y = newsample$Cactus
m = newsample$m
N = 900
n1 = 40
m_threshold = 7
new_y_HT(
	y = newsample$Cactus, 
	N = 900, 
	n1 = 40, 
	m = newsample$m, 
	m_threshold = 7
)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}

