% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_i.R
\name{pi_i}
\alias{pi_i}
\title{Calculate the inclusion probability of unit $i$ in a simple random sample without replacement.}
\usage{
pi_i(N, n1, m)
}
\arguments{
\item{N}{Population size.}

\item{n1}{Initial sample size.}

\item{m}{Vector of values giving the number of units satisfying the ACS criterion in network $i$.}
}
\value{
vector of inclusion probabilities
}
\description{
Calculate the inclusion probability of unit $i$ in a simple random sample without replacement.
}
\examples{
# example from section 4, p. 1055 (first column) in Thompson (1990)
N = 5
n1 = 2
y = c(1,10,1000)
m = c(1,2,2)
pi_i(N, n1, m)
# EXAMPLE 4:
# Ch. 24, Exercise #2, p. 307, from Thompson (2002)
N=1000
n1=100
m=c(2,3,rep(1,98))
head(pi_i(N, n1, m))
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.

Thompson. 1990.
}
