% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_i.R
\name{pi_i_replace}
\alias{pi_i_replace}
\title{Calculate the inclusion probabilities of units in a simple random sample with replacement.}
\usage{
pi_i_replace(N, n1, m_vec)
}
\value{
vector of inclusion probabilities
}
\description{
Calculate the inclusion probabilities of units in a simple random sample with replacement.
}
