% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_i.R
\name{pi_i_replace}
\alias{pi_i_replace}
\title{Calculate the inclusion probabilities of units in a simple random sample with replacement.}
\usage{
pi_i_replace(N, n1, m_vec)
}
\arguments{
\item{N}{Population size.}

\item{n1}{An integer giving the initial sample size (ie., the size of the initial simple random sample).}

\item{m_vec}{Vector of values \eqn{m} for the set of units in a sample, of length \eqn{n1}. Each \eqn{m} value within the vector \code{m_vec} denotes the number of units satisfying the ACS criterion for the network \eqn{i} to which the unit belongs.}
}
\value{
vector of inclusion probabilities
}
\description{
Calculate the inclusion probabilities of units in a simple random sample with replacement.
}
