% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ij.R
\name{pi_ij}
\alias{pi_ij}
\title{Calculate joint inclusion probability of unit $j$ and $h$}
\usage{
pi_ij(N, n1, m_vec)
}
\arguments{
\item{N}{Population size.}

\item{n1}{An integer giving the initial sample size (ie., the size of the initial simple random sample).}

\item{m_vec}{Vector of values $m$ for the set of units in a sample, of length $n1$. Each $m$ value within the vector \code{m_vec} denotes the number of units satisfying the ACS criterion for the network $i$ to which the unit belongs.}
}
\description{
Calculate joint inclusion probability of unit $j$ and $h$

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}
\examples{
# Thompson sampling book, ch. 24 exercises, p. 307, number 2
library(magrittr)
N=1000
n1=100
m=c(2,3,rep(1,98))
pi_ij(N, n1, m) \%>\% .[1,2]
}
