% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_ij.R
\name{pi_ij_RACS}
\alias{pi_ij_RACS}
\title{Calculate joint inclusion probability of unit $j$ and $h$, using the RACS correction}
\usage{
pi_ij_RACS(N, n1, m_vec, m_threshold)
}
\arguments{
\item{m_threshold}{threshold value above which to calculate pi_i and pi_j differently.}
}
\description{
Calculate joint inclusion probability of unit $j$ and $h$, using the RACS correction

Sauby, K.E and Christman, M.C. \emph{In preparation.} A Sampling Strategy Designed to Maximize the Efficiency of Data Collection of Food Web Relationships
}
