% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSimulationResults.R
\name{plotSimulationResults}
\alias{plotSimulationResults}
\title{Plot Simulation Results}
\usage{
plotSimulationResults(dataset, variable = NULL, x_variable = NULL,
  y_variable = NULL, grouping_variable, legend_name, xlab_name = NULL,
  ylab_name = NULL, SD = FALSE, facet = facet,
  legendposition = "bottom", legendcolor = "black", x_angle = 360,
  xlength = 6, ylength = 4, roundn = 2, y_breaks = NULL)
}
\arguments{
\item{dataset}{Simulation data on sampling of the multiple patch realizations.}

\item{variable}{The variable to be plotted. Default value is \code{NULL}.}

\item{x_variable}{The variable to be plotted on the x-axis. Default value is \code{NULL}.}

\item{y_variable}{The variable to be plotted to be plotted on the y-axis. Default value is \code{NULL}.}

\item{grouping_variable}{The variable dividing the data into subgroups.}

\item{legend_name}{The name given to the \code{grouping_variable} in the legend.}

\item{xlab_name}{The x-axis name. Default value is \code{NULL}.}

\item{ylab_name}{The y-axis name. Default value is \code{NULL}.}

\item{roundn}{}

\item{SD=FALSE}{}

\item{facet=facet}{}

\item{legendposition="bottom"}{}

\item{legendcolor="black"}{}

\item{x_breaks}{}

\item{nlength}{Number of breaks on the x-axis. Default value is \code{6}.}
}
\value{
Plot of the true variable parameter values (x-axis) versus variable bias (y-axis). The plot is divided into three facets: Horvitz-Thompson estimators, SRSWOR estimators (excluding information from adaptively sampled plots), and SRSWOR estimators applied to the entire dataset (ignoring the sampling design). Error bars indicate the standard deviation.
}
\description{
Plot Simulation Results
}
