% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_hat.R
\name{R_hat}
\alias{R_hat}
\title{Calculate the Horvitz-Thompson Ratio Estimator}
\usage{
R_hat(y, x, N, n1, m, with_replacement = "FALSE")
}
\arguments{
\item{y}{Attribute data about species of interest y (e.g., abundance, presence/absence).}

\item{x}{Auxiliary data about object of interest y.}

\item{N}{Population size.}

\item{n1}{Initial sample size.}

\item{m}{Number of units satisfying the ACS criterion in network $i$.}

\item{with_replacement}{Whether sampling should be done with or without replacement. Defaults to \code{FALSE}.}
}
\description{
R_hat is calculated by dividing the Horvitz-Thompson estimator, Tau_hat_z by Tau_hat_x. See p. 77, Thompson (2002). THERE CANNOT BE NAS IN X OR Y?
}
\examples{
# Example from Thompson (2002), p. 78-79
N = 100
n1 = 4
y = c(60, 14, 1)
x = c(1, 1, 1)
m = c(5, 2, 1)
R_hat(y, x, N, n1, m, with_replacement="TRUE")
var_R_hat(y, x, N, n1, m, with_replacement="TRUE")
}
