% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleGridPopulation.R
\name{sampleGridPop}
\alias{sampleGridPop}
\title{Select a sample of cells from the grid population and randomly assign species information.}
\usage{
sampleGridPop(grid, n.networks, cluster.centers, seed)
}
\arguments{
\item{grid}{A dataframe of x and y coordinates (can be created with \code{createPop}).}

\item{n.networks}{Initial sample size that determines the relative density of the species.}

\item{cluster.centers}{Dataframe including x and y coordinates and network IDs for the centers of clusters.}

\item{seed}{Vector of numbers to be given to \code{set.seed()}. Two numbers are used: the first to determine the sample of locations and the second to determine the sample of species information to be assigned to the sample locations.}
}
\description{
This function uses two random numbers: the first to set the seed to sample locations from a grid, and the second to sample networks to assign to those locations (more specifically, the centers of the networks are assigned to those locations).
}
\examples{
library(magrittr)
library(dplyr)
grid = createPop(0, 25, 0, 25)
n.networks = 2
data(Thompson1990Fig1Pop)
cluster.info = Thompson1990Fig1Pop \%>\% 
	filter(m > 1) \%>\%
	createNetworks
cluster.centers = filter(cluster.info, Rel_x==0 & Rel_y==0) \%>\%
dplyr::select(-c(x,y))
seed = 1:2
sampleGridPopulation(grid, n.networks, cluster.centers, seed)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
