% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizeClusters.R
\name{sampleGridPop}
\alias{sampleGridPop}
\title{Assign a sample of cluster centers (derived from real data) to a sample of grid locations}
\usage{
sampleGridPop(grid, n.networks, NetworkIDs, seed)
}
\arguments{
\item{grid}{A dataframe of x and y coordinates (can be created with \code{createPop}).}

\item{n.networks}{The number of networks to be placed in the realization.}

\item{NetworkIDs}{Vector of NetworkIDs to randomly assign to the grid.}

\item{seed}{Vector of size two of numbers to be given to \code{set.seed()}. The first number is given to the function \code{sample} to create a sample of locations and the second number is given to the function \code{sample} to randomly assign the NetworkIDs to the sample of locations.}
}
\description{
This function uses two random numbers: the first to set the seed to sample locations from a grid, and the second to determine the NetworkIDs to be assigned to those locations (more specifically, the centers of the networks are assigned to those locations).
}
\examples{
library(magrittr)
library(dplyr)
grid = createPop(0, 25, 0, 25)
n.networks = 2
data(Thompson1990Fig1Pop)
NetworkIDs = Thompson1990Fig1Pop \%>\% 
	filter(m > 1) \%$\%
	unique(NetworkID)
seed = 1:2
sampleGridPop(grid, n.networks, NetworkIDs, seed)
}
