% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleRealizations.R
\name{sampleRealizations}
\alias{sampleRealizations}
\title{Sample species patch realizations simulations}
\usage{
sampleRealizations(
  popdata,
  sims,
  n1_vec,
  avar = NULL,
  ovar,
  rvar = NULL,
  SamplingDesign = "ACS",
  yvar,
  y_HT_formula = "y_HT",
  var_formula = "var_y_HT",
  m_threshold = NULL,
  f_max = 2,
  SampleEstimators = FALSE,
  SpatStat = TRUE,
  mChar = TRUE,
  popvar,
  realvar = "realization",
  weights = "S",
  seed = NA
)
}
\arguments{
\item{popdata}{patch realizations}

\item{sims}{Number of simulations per population.}

\item{n1_vec}{Vector of initial sample size(s) for the initial simple random sample(s) without replacement; can be a single value or vector of values.}

\item{avar}{The total vector of variables (\code{c(avar, ovar, rvar)}) should be a length of at least 1.}

\item{ovar}{Vector of occupancy variables. The total vector of variables (\code{c(avar, ovar, rvar)}) should be a length of at least 1.}

\item{rvar}{Vector of ratio variables. The total vector of variables (\code{c(avar, ovar, rvar)}) should be a length of at least 1.}

\item{SamplingDesign}{A character string supplying the sampling design to use; accepted options included "SRS" (simple random sampling), "ACS" (unrestricted adaptive cluster sampling), and "RACS" (restricted ACS). The default is "ACS".}

\item{yvar}{A string giving the name of the variable of interest, \eqn{y}, in the supplied dataframe \eqn{popdata}. This variable determines the condition under which adaptive cluster sampling takes place. In the dataframe $popdata$ this variable \eqn{y} must be numeric.}

\item{y_HT_formula}{The formula used to estimate the population total: 
either the Horvitz-Thompson estimator, 'y_HT,' or or the RACS-corrected 
Horvitz-Thompson estimator, 'y_HT_RACS'.}

\item{var_formula}{The formula used to estimate the variance of the 
population total: either the Horvitz-Thompson variance estimator, 'var_y_HT',
or the RACS-corrected Horvitz-Thompson variance estimator, "var_y_HT_RACS." 
Defaults to "var_y_HT".}

\item{m_threshold}{threshold value above which to calculate \eqn{pi_i} and \eqn{pi_j} differently.}

\item{f_max}{The maximum number of expanding sets of adjacent units surveyed
per primary unit in the adaptive phase of ACS. For more information, see 
Sauby and Christman (in prep).}

\item{SampleEstimators}{If "TRUE", calculate the sample mean and sample 
variance for each simulation. Default is FALSE.}

\item{SpatStat}{TRUE or FALSE. If "TRUE", for each simulation calculate 
Moran's I, and the nugget, sill, and range of the semivariogram. Default is 
TRUE.}

\item{mChar}{TRUE or FALSE. If "TRUE", for each simulation calculate summary 
statistics (median, mean, min, and max) for the sample's m values. Also, for 
each simulation and for the set of unique m values, calculate the same 
summary statistics. If "FALSE," no summary statistics are calculated.}

\item{popvar}{Categorical variable used to identify different populations.}

\item{realvar}{Variable identifying each realization. Default is "realization"}

\item{weights}{If SpatStat is "TRUE", this is a vector giving 
spatial weight matrix styles to use to calculate the Join Count and Moran's I
 statistics. Can take on values "W", "B", "C", "U", "S", and "minmax". See 
 \code{nb2listw} for more details.}

\item{seed}{A vector of numbers, equal in length to n1_vec, to set random seeds, if a goal is the ability to reproduce the random sampling.}
}
\description{
This function simulates sampling of multiple realizations of 
patches of the species of interest within the grid of locations created with 
\code{createPop}. The number of total simulations is
length(n1_vec) x length(popvar) x length(realvar)
}
\examples{
# sims=20
# n1_vec=c(5,10,20,40)
# population <- createPop(x_start = 1, x_end = 30, y_start = 1, y_end = 30)
# #' avar = NULL
# ovar = c(
# 	"Stricta",
# 	"Pusilla",
#  "Cactus",
#  "CACA_on_Pusilla",
# "CACA_on_Stricta",
# "MEPR_on_Pusilla",
# "MEPR_on_Stricta",
# "Old_Moth_Evidence_Pusilla",
# "Old_Moth_Evidence_Stricta"
# )
# data(CactusRealizations)
# popdata = CactusRealizations # WHY IS THERE ISLAND=NA
# simulation_data <- sampleRealizations(
# popdata = popdata,
# 	sims = sims,
# 	n1_vec = n1_vec,
# 	avar = avar,
# 	ovar = ovar,
# popvar="Island",
# yvar="Cactus"
# )
# sims=200
# n1_vec=c(75,150,225,300,350)
# simulation_data_SRSWOR <- sampleRealizations(
# popdata = popdata,
# sims = sims,
# n1_vec = n1_vec,
# avar = avar,
# ovar = ovar,
# popvar="Island"
# )
}
\references{
\insertRef{saubyadaptive}{ACSampling}
}
