% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_data.R
\name{simulation_data}
\alias{simulation_data}
\title{Simulation Data}
\format{A data frame with 400 rows and 5 variables}
\description{
Realizations created from my own data.
}
\details{
\itemize{
  \item variable List variables here
}
}
\examples{
# include example here
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.

Thompson, S. (1990). Adaptive Cluster Sampling. \emph{Journal of the American Statistical Association}, 85(412): 1050--1059.
}
