% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Thompson1990Figure1Sample.R
\name{Thompson1990Fig1Sample}
\alias{Thompson1990Fig1Sample}
\title{Sample from Figure 1 from Thompson (1990)}
\format{
A data frame with 10 rows and 2 variables
}
\description{
An example dataset from Figure 1 from Thompson (1990) containing the x- and y-coordinates of units selected in a simple random sample without replacement.

\itemize{
  \item x. X coordinate.
  \item y. Y coordinate.
}
}
\references{
\insertRef{thompson1990adaptive}{ACSampling}
}
