% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{var_Hajek}
\alias{var_Hajek}
\title{Variance estimator free of joint inclusion probability calculations for unequal probability sampling, Hajek}
\usage{
var_Hajek(n, y, pi_i_values)
}
\arguments{
\item{n}{sample size (an integer value).}

\item{y}{Vector of values observed for the variable of interest, $y$.}

\item{pi_i_values}{vector of first-order inclusion probabilities, calculated using \code{pi_i}.}
}
\description{
Variance estimator free of joint inclusion probability calculations for unequal probability sampling, Hajek
}
\references{
Berger, Y. G. (2005). Variance estimation with Chao's sampling scheme. Journal of Statistical Planning and Inference, 127(1-2), 253-277. http://doi.org/10.1016/j.jspi.2003.08.014
}
