% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variance_wo_jointinclusionprobabilities.R
\name{var_pi}
\alias{var_pi}
\title{Variance estimator free of joint inclusion probability calculations for unequal probability sampling}
\usage{
var_pi(n, y, pi_i_values, estimator)
}
\arguments{
\item{n}{sample size (an integer value).}

\item{y}{need description}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}

\item{estimator}{Options include "Hajek".}
}
\description{
This gives equation 9 on page 10 in Berger and Tille 2009.
}
\examples{
# Hajek Approximation
library(dplyr)
library(magrittr)
Z = createACS(Thompson1990Fig1Pop, seed=3, n1=30, "y_value", condition=0)
Z_summary <- Z \%>\% 
	dplyr::filter(Sampling!="Edge") \%>\%
	group_by(NetworkID) \%>\%
	filter(NetworkID > 0) \%>\%
	dplyr::summarise(
		m = m[1],
		y_total = sum(y_value, na.rm=TRUE)
	)
var_y_HT(
	N = dim(Thompson1990Fig1Pop)[1], 
	n1 = dim(Thompson1990Figure1Sample)[1], 
	m = Z_summary$m, 
	y = Z_summary$y_total
)
pi_i_values <- pi_i(N=900,n1=30, m=Z_summary$m)
Hajek(pi_i=pi_i_values, n=30)
var_pi(
	n = 30, 
	y = Z_summary$y_total, 
	pi_i_values = pi_i_values,
	estimator = "Hajek"
)
}
\references{


Berger, Y. G., & Tille, Y. (2009). Sampling with Unequal Probabilities. Handbook of Statistics, 1-17.
#' Hajek, J. (1964). Asymptotic theory of rejective sampling with varying probabilities from a finite population. The Annals of Mathematical Statistics.
}
