% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_R_hat_C.R
\name{var_R_hat}
\alias{var_R_hat}
\title{Calculate the variance of R hat}
\usage{
var_R_hat(y, x, N, n1, m_vec, replace = "FALSE")
}
\arguments{
\item{y}{The variable of interest, $y$. Must be a numeric vector. The criterion that determines whether adaptive cluster sampling takes place is based on this variable.}

\item{x}{A vector of values for variable $x$, which is auxiliary data about the variable of interest $y$.}

\item{replace}{Whether sampling should be done with or without replacement. Defaults to \code{FALSE}.}
}
\description{
Calculate the variance of R hat
}
\examples{
@template ex_Thompson2002_p_78
}
