% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_y_HT_C.R
\name{var_y_HT_RACS}
\alias{var_y_HT_RACS}
\title{Calculate the variance of the Horvitz-Thompson estimator of the mean using the RACS correction}
\usage{
var_y_HT_RACS(N, n1, m, y, m_threshold, pi_i_values = NULL)
}
\arguments{
\item{N}{Population size}

\item{n1}{Initial sample size}

\item{m}{Vector of $m$, each corresponding to a unique network.}

\item{y}{Vector of $y$ total, each corresponding to a unique network.}

\item{m_threshold}{threshold value above which to calculate pi_i and pi_j differently.}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}
}
\description{
Calculate the variance of the Horvitz-Thompson estimator of the mean using the RACS correction
}
\examples{
library(ggplot2)
library(magrittr)
library(dplyr)
# Sampling of population from Figure 1, Thompson (1990)

data(Thompson1990Fig1Pop)
data(Thompson1990Figure1Sample)

# plot sample overlaid onto population
ggplot() +
	geom_point(data=Thompson1990Fig1Pop, aes(x,y, 
	size=factor(y_value),
		shape=factor(y_value))) +
	scale_shape_manual(values=c(1, rep(16, length(2:13)))) +
	geom_point(data=Thompson1990Figure1Sample, aes(x,y), shape=0, size=7)

# INITIATE ACS
Z = createACS(population_data=Thompson1990Fig1Pop, 
	n1=dim(Thompson1990Figure1Sample)[1], 
	initial_sample=Thompson1990Figure1Sample, y_variable="y_value")

# CALCULATE var(y_HT)
# create dataframe of network info
Z_summary <- Z \%>\% group_by(NetworkID) \%>\%
	summarise(
		m = m[1],
		y_total = sum(y_value, rm.na=TRUE)
		) \%>\%
		filter(NetworkID > 0)

var_y_HT_RACS(
	N = dim(Thompson1990Fig1Pop)[1], 
	n1 = dim(Thompson1990Figure1Sample)[1], 
	m = Z_summary$m, 
	y = Z_summary$y_total,
	m_threshold=3
)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
