% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_y_HT_C.R
\name{var_y_HT_RACS}
\alias{var_y_HT_RACS}
\title{Calculate the variance of the Horvitz-Thompson estimator of the mean using the RACS correction}
\usage{
var_y_HT_RACS(N, n1, m, y, m_threshold, pi_i_values = NULL)
}
\arguments{
\item{N}{Population size}

\item{n1}{Initial sample size}

\item{m}{Vector of $m$, each corresponding to a unique network.}

\item{y}{Vector of $y$ total, each corresponding to a unique network.}

\item{m_threshold}{threshold value above which to calculate pi_i and pi_j differently.}

\item{pi_i_values}{vector of inclusion probabilities, if not calculated using this function. Default is \code{NULL}.}
}
\description{
Calculate the variance of the Horvitz-Thompson estimator of the mean using the RACS correction
}
\examples{
patch = patch_data \%>\% filter(n.networks==unique(patch_data$n.networks)[3])
S <- createSRS(population_data=patch, seed=26, n1=10)
initial_sample <- S[, c("x", "y")]
RACS = createRACS(
	population_data=patch, 
	seed=26, 
	n1=10, 
	y_variable="Cactus", 
	initial_sample=initial_sample
) \%>\% mutate(Design="(b) RACS, n = 45")
var_y_HT_RACS(N=dim(patch)[1], n1=10, m=RACS$m, y=RACS$Cactus, m_threshold=7)
}
\references{
Sauby, K.E and Christman, M.C. \emph{In preparation.} Restricted adaptive cluster sampling.
}
